% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{AbstractPolicy}
\alias{AbstractPolicy}
\title{External AbstractPolicy}
\description{
AbstractPolicy intro
}
\section{Usage}{

\preformatted{b <- AbstractPolicy$new()

b$reset()

print(b)
}
}

\section{Arguments}{

\describe{
\item{b}{A \code{AbstractPolicy} object.}
}
}

\section{Details}{

\code{$new()} starts a new AbstractPolicy,
it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.
}

\examples{
\dontrun{}

}
