% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{Agent}
\alias{Agent}
\title{External Agent}
\description{
Agent intro
}
\section{Usage}{

\preformatted{b <- Agent$new()

b$reset()

print(b)
}
}

\section{Arguments}{

\describe{
\item{b}{A \code{Agent} object.}
}
}

\section{Details}{

\code{$new()} starts a new Agent, it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.
}

\examples{
\dontrun{}

}
