% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.basic.R
\name{BasicAgent}
\alias{BasicAgent}
\title{External BasicAgent}
\description{
BasicAgent intro
}
\section{Usage}{

\preformatted{b <- BasicAgent$new()

b$reset()

print(b)
}
}

\section{Arguments}{

\describe{
  \item{b}{A \code{BasicAgent} object.}
}
}

\section{Details}{

\code{$new()} starts a new BasicAgent, it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.
}

\examples{
\dontrun{}

}
