% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandit.R, R/bandit_basic.R
\name{BasicBandit}
\alias{BasicBandit}
\alias{BasicBandit}
\title{External BasicBandit}
\description{
BasicBandit intro

BasicBandit intro
}
\section{Usage}{

\preformatted{b <- BasicBandit$new()

b$reset()

print(b)
}


\preformatted{b <- BasicBandit$new()

b$reset()

print(b)
}
}

\section{Arguments}{

\describe{
  \item{b}{A \code{BasicBandit} object.}
}


\describe{
  \item{b}{A \code{BasicBandit} object.}
}
}

\section{Details}{

\code{$new()} starts a new BasicBandit, it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.


\code{$new()} starts a new BasicBandit, it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.
}

\examples{
\dontrun{}

\dontrun{}

}
