% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_mab_btslogit.R
\name{BTSLogitPolicy}
\alias{BTSLogitPolicy}
\title{Policy: BTSLogitPolicy}
\description{
Each time step t, \code{BTS} runs a linear regression that produces coefficients for each context feature \code{d}.
It then observes the new context, and generates a predicted payoff or reward together with a confidence interval for each available arm.
It then proceeds to choose the arm with the highest upper confidence bound.
}
\section{Usage}{

\preformatted{
policy <- BTSLogitPolicy(alpha = 1.0, name = "LinUCB")
}
}

\seealso{
Other contextual policies: \code{\link{AbstractPolicy}},
  \code{\link{ContextualEpsilonGreedyPolicy}},
  \code{\link{ContextualThompsonSamplingPolicy}},
  \code{\link{EpsilonFirstPolicy}},
  \code{\link{EpsilonGreedyPolicy}},
  \code{\link{Exp3Policy}},
  \code{\link{LinUCBDisjointPolicy}},
  \code{\link{LinUCBHybridPolicy}},
  \code{\link{OraclePolicy}}, \code{\link{RandomPolicy}},
  \code{\link{SimpleBTSPolicy}},
  \code{\link{SoftmaxPolicy}},
  \code{\link{ThompsonSamplingPolicy}},
  \code{\link{UCB1Policy}}
}
