% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-output.R
\name{check.genmatrix}
\alias{check.genmatrix}
\title{Check that A Genmatrix is Compatible with a Configuration}
\usage{
check.genmatrix(config, genmatrix)
}
\value{
Returns TRUE if successful; else throws an error.
}
\description{
Check that A Genmatrix is Compatible with a Configuration
}
