% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{ci_boot}
\alias{ci_boot}
\title{Bootstrapped confidence interval}
\usage{
ci_boot(x, conf_int = 0.95, B = 1000, na_rm = TRUE, reps = FALSE)
}
\arguments{
\item{x}{a numeric vector from which NAs will be removed automatically}

\item{conf_int}{specifies the confidence level (0-1) for interval estimation of
the population mean.}

\item{B}{number of bootstrap resamples}

\item{na_rm}{defaults to \code{TRUE} unlike built-in functions, so that by default
\code{NA}s are automatically removed.}

\item{reps}{set to \code{TRUE} to have \code{smean.cl.boot} return the vector of
bootstrapped means as the \code{reps} attribute of the returned object.}
}
\value{
A vector of summary statistics.
}
\description{
From HMisc. A very fast implementation of the basic nonparametric bootstrap for
obtaining confidence limits for the population mean without assuming normality.
}
\author{
Frank Harrell
}
