% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{clipr}
\alias{clipr}
\title{Clip vectors}
\usage{
clipr(x, min, max)
}
\arguments{
\item{x}{to be clipped vector}

\item{min}{numeric. lowest value}

\item{max}{numeric. highest value}
}
\description{
Clips values to a mininum and maximum value. That is, all values below the lower clamp
value and the upper clamp value become the lower/upper value specified
}
