% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{collapsepatmatrix}
\alias{collapsepatmatrix}
\title{Construct a Pattern-Substring Matrix}
\usage{
collapsepatmatrix(ipatterns, leftwin, shortwin = nchar(fpatterns[1]),
  rightwin = nchar(ipatterns[1]) - shortwin - leftwin,
  fpatterns = getpatterns(nchar(ipatterns[1]) - leftwin - rightwin, bases),
  bases)
}
\value{
A (nbases)^k by (nbases)^{k-leftwin-rightwin} matrix projection matrix that
   maps patterns onto the shorter patterns obtained by deleting leftwin
   characters at the start and rightwin characters at the end.
}
\description{
Construct the matrix U described in the tex.
ipatterns are the "input" patterns, while fpatterns are the "final" projected patterns
This function assumes that all input patterns are the same length.
}
