% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{computeresids}
\alias{computeresids}
\title{Compute Residuals}
\usage{
computeresids(model, pretty = TRUE, longwin = longwin(model),
  shortwin = shortwin(model), leftwin = leftwin(model), counts = NULL,
  ...)
}
\arguments{
\item{model}{An object inheriting from `context` or `contextTree` having a `fitted` method.}

\item{pretty}{Whether to return results in a data frame sorted by z-score (otherwise, returns a matrix of the same form as `counts`).}

\item{longwin}{The `long` window length of Tmers to compute residuals for.}

\item{shortwin}{The `short` window length of Tmers to compute residuals for.}

\item{leftwin}{The left overhang of Tmers to compute residuals for.}

\item{counts}{A tuplecounts object - can be supplied to compute residuals for longer Tmers than the model was fit with.}

\item{...}{Additional arguments to be passed to `fitted()`.}
}
\value{
If `pretty` is TRUE, returns a data frame with columns `inpat` (the head of the Tmer), `outpat` (the tail of the Tmer), 
`observed`, `expected`, `resid` (observed minus expected), and `z` (residual divided by square root of expected).
Otherwise, returns a matrix of the same dimensions as `counts` with rows and columns labeled by the head and tail Tmer patterns.
}
\description{
Compute Residuals
}
