% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-output.R
\name{config.dereference}
\alias{config.dereference}
\title{Track down references in a config file.}
\usage{
config.dereference(config, x)
}
\description{
Follow pointers in config: if config[[x]] is a string, it refers to another entry.
}
