% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_classes.R
\docType{methods}
\name{dimnames,context-method}
\alias{dimnames,context-method}
\alias{counts,context-method}
\alias{countframe,context-method}
\alias{longwin}
\alias{shortwin}
\alias{leftwin}
\alias{rightwin}
\alias{longwin,context-method}
\alias{shortwin,context-method}
\alias{leftwin,context-method}
\alias{nmuts}
\alias{nsel}
\alias{fixparams}
\alias{coef,context-method}
\alias{rowSums,context-method}
\alias{fitted,context-method}
\alias{residuals,context-method}
\title{Methods to get window lengths.}
\usage{
\S4method{dimnames}{context}(x)

\S4method{counts}{context}(x)

\S4method{countframe}{context}(x)

longwin(x)

shortwin(x)

leftwin(x)

rightwin(x)

\S4method{longwin}{context}(x)

\S4method{shortwin}{context}(x)

\S4method{leftwin}{context}(x)

nmuts(x)

nsel(x)

fixparams(x)

\S4method{coef}{context}(object)

\S4method{rowSums}{context}(x)

\S4method{fitted}{context}(object, ...)

\S4method{residuals}{context}(object, ...)
}
\description{
Methods to get window lengths.
}
