% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_classes.R
\docType{methods}
\name{nmuts,contextModel-method}
\alias{nmuts,contextModel-method}
\alias{nsel,contextModel-method}
\alias{fixparams,contextModel-method}
\alias{coef,contextModel-method}
\title{Methods for contextModel objects}
\usage{
\S4method{nmuts}{contextModel}(x)

\S4method{nsel}{contextModel}(x)

\S4method{fixparams}{contextModel}(x)

\S4method{coef}{contextModel}(object)
}
\description{
Methods for contextModel objects
}
