% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_classes.R
\docType{methods}
\name{dimnames,contextTree-method}
\alias{dimnames,contextTree-method}
\alias{counts,contextTree-method}
\alias{countframe,contextTree-method}
\alias{longwin,contextTree-method}
\alias{shortwin,contextTree-method}
\alias{leftwin,contextTree-method}
\alias{nmuts,contextTree-method}
\alias{nsel,contextTree-method}
\alias{fixparams,contextTree-method}
\alias{coef,contextTree-method}
\alias{rowSums,contextTree-method}
\title{Methods for contextTree}
\usage{
\S4method{dimnames}{contextTree}(x)

\S4method{counts}{contextTree}(x)

\S4method{countframe}{contextTree}(x)

\S4method{longwin}{contextTree}(x)

\S4method{shortwin}{contextTree}(x)

\S4method{leftwin}{contextTree}(x)

\S4method{nmuts}{contextTree}(x)

\S4method{nsel}{contextTree}(x)

\S4method{fixparams}{contextTree}(x)

\S4method{coef}{contextTree}(object)

\S4method{rowSums}{contextTree}(x)
}
\description{
Methods for contextTree
}
