% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contextual.R
\name{Contextual}
\alias{Contextual}
\title{External Contextual}
\description{
Contextual intro
}
\section{Usage}{

\preformatted{b <- Contextual$new()

b$reset()

print(b)
}
}

\section{Arguments}{

\describe{
  \item{b}{A \code{Contextual} object.}
}
}

\section{Details}{

\code{$new()} starts a new Contextual, it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.
}

\examples{
\dontrun{}

}
