% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandit_contextual.R
\name{ContextualBandit}
\alias{ContextualBandit}
\title{Bandit: ContextualBandit}
\description{
ContextualBandit intro
}
\section{Usage}{

\preformatted{b <- ContextualBandit$new()

b$reset()

print(b)
}
}

\section{Arguments}{

\describe{
\item{b}{A \code{ContextualBandit} object.}
}
}

\section{Details}{

\code{$new()} starts a new ContextualBandit, it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.
}

\seealso{
Core contextual classes: \code{\link{Bandit}}, \code{\link{Policy}}, \code{\link{Simulator}},
\code{\link{Agent}}, \code{\link{History}}, \code{\link{Plot}}

Other contextual subclasses: \code{\link{BasicBandit}},
  \code{\link{ContextualDisjointThompsonSamplingPolicy}},
  \code{\link{ContextualThompsonSamplingPolicy}},
  \code{\link{ContinuumBandit}},
  \code{\link{DoublyRobustOfflineBandit}},
  \code{\link{EpsilonFirstPolicy}},
  \code{\link{EpsilonGreedyPolicy}},
  \code{\link{Exp3Policy}},
  \code{\link{GittinsBrezziLaiPolicy}},
  \code{\link{LiSamplingOfflineBandit}},
  \code{\link{LifPolicy}},
  \code{\link{LinUCBDisjointPolicy}},
  \code{\link{LinUCBDisjointSmPolicy}},
  \code{\link{LinUCBHybridPolicy}},
  \code{\link{OraclePolicy}}, \code{\link{RandomPolicy}},
  \code{\link{SoftmaxPolicy}},
  \code{\link{SyntheticBandit}},
  \code{\link{ThompsonSamplingPolicy}},
  \code{\link{UCB1Policy}}
}
