% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandit_cmab_bernoulli.R
\name{ContextualBernoulliBandit}
\alias{ContextualBernoulliBandit}
\title{Bandit: ContextualBernoulliBandit}
\description{
Contextual extension of \code{\link{BasicBernoulliBandit}}.
}
\details{
Contextual extension of \code{\link{BasicBernoulliBandit}} where a user specified \code{d x k} dimensional
matrix takes the place of \code{\link{BasicBernoulliBandit}} \code{k} dimensional probability vector. Here,
each row \code{d} represents a feature with \code{k} reward probability values per arm.

For every \code{t}, \code{ContextualBernoulliBandit} randomly samples from its \code{d} features/rows at
random, yielding a binary \code{context} matrix representing sampled (all 1 rows) and unsampled (all 0)
features/rows. Next, \code{ContextualBernoulliBandit} generates \code{rewards} contingent on either sum or
mean (default) probabilities of each arm/column over all of the sampled features/rows.

Illustrates precaching of contexts and rewards.
}
\section{Usage}{

\preformatted{
  bandit <- ContextualBernoulliBandit$new(weights = NULL, precaching  = TRUE, sum_weights = FALSE)
}
}

\section{Arguments}{


\describe{
\item{\code{weights}}{
numeric matrix; \code{d x k} dimensional matrix where each row \code{d} represents a feature with
\code{k} reward probability values per arm.
}
\item{\code{precaching}}{
logical; determines if the bandit precaches all contexts and rewards.
}
\item{\code{sum_weights}}{
logical; determines whether \code{ContextualBernoulliBandit} takes the sum or the mean (default) of
an arm's probability values over all selected feature rows.
}
}
}

\section{Methods}{


\describe{

\item{\code{new(weights = NULL, precaching  = TRUE, sum_weights = FALSE)}}{
generates and instantializes a new \code{Bandit} instance.
For arguments, see Argument section above.
}

\item{\code{get_context(t)}}{
argument:
\itemize{
\item \code{t}: integer, time step \code{t}.
}
returns a named \code{list}
containing the current \code{d x k} dimensional matrix \code{context$X},
number of arms \code{context$k} and number of features \code{context$d}.
}

\item{\code{get_reward(t, context, action)}}{
arguments:
\itemize{
\item \code{t}: integer, time step \code{t}.
\item \code{context}: list, with \code{context$k} (number of arms).
\item \code{action}:  list, containing \code{action$choice} (as set by \code{policy}).
}
returns a named \code{list} containing \code{reward$reward}
}

\item{\code{generate_bandit_data()}}{
helper function called before \code{Simulator} starts iterating over all time steps \code{t} in T.
This function is only called when \code{bandit$precaching} is \code{TRUE} (default \code{FALSE}).
Pregenerates \code{contexts} and \code{rewards}.
}
}
}

\examples{
\donttest{
horizon            <- 100L
simulations        <- 100L

# rows represent features, columns represent arms:

context_weights    <- matrix(  c(0.4, 0.2, 0.4,
                                 0.3, 0.4, 0.3,
                                 0.1, 0.8, 0.1),  nrow = 3, ncol = 3, byrow = TRUE)

bandit             <- ContextualBernoulliBandit$new(weights = context_weights, precaching = FALSE)

agents             <- list( Agent$new(EpsilonGreedyPolicy$new(0.1), bandit),
                            Agent$new(LinUCBDisjointOptimizedPolicy$new(0.6), bandit))

simulation         <- Simulator$new(agents, horizon, simulations, do_parallel = FALSE)
history            <- simulation$run()

plot(history, type = "cumulative")

}
}
\seealso{
Core contextual classes: \code{\link{Bandit}}, \code{\link{Policy}}, \code{\link{Simulator}},
\code{\link{Agent}}, \code{\link{History}}, \code{\link{Plot}}

Bandit subclass examples: \code{\link{BasicBernoulliBandit}}, \code{\link{ContextualLogitBandit}}, \code{\link{OfflinePolicyEvaluatorBandit}}

Policy subclass examples: \code{\link{EpsilonGreedyPolicy}}, \code{\link{ContextualThompsonSamplingPolicy}}
}
