% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_cmab_ts_disjoint.R
\name{ContextualDisjointThompsonSamplingPolicy}
\alias{ContextualDisjointThompsonSamplingPolicy}
\title{Policy: Contextual Thompson Sampling with Linear Payoffs}
\description{
\code{ContextualDisjointThompsonSamplingPolicy} works by maintaining a prior on the the mu_hat rewards of its arms.
In this, it follows a beta–binomial model with parameters alpha and beta, sampling values
for each arm from its prior and picking the arm with the highest value.
When an arm is pulled and a Bernoulli reward is observed, it modifies the prior based on the reward.
This procedure is repeated for the next arm pull.
}
\section{Usage}{

\preformatted{
policy <- ContextualDisjointThompsonSamplingPolicy(alpha = 1, beta = 1)
}
}

\section{Arguments}{


\describe{
\item{\code{alpha}}{
integer, a natural number N>0 - first parameter of the Beta distribution
}
\item{\code{beta}}{
integer, a natural number N>0 - second parameter of the Beta distribution
}
\item{\code{name}}{
character string specifying this policy. \code{name}
is, amongst others, saved to the History log and displayed in summaries and plots.
}
}
}

\section{Methods}{


\describe{
\item{\code{new(alpha = 1, beta = 1)}}{ Generates a new \code{ContextualDisjointThompsonSamplingPolicy} object. Arguments are defined in the Argument section above.}
}

\describe{
\item{\code{set_parameters()}}{each policy needs to assign the parameters it wants to keep track of
to list \code{self$theta_to_arms} that has to be defined in \code{set_parameters()}'s body.
The parameters defined here can later be accessed by arm index in the following way:
\code{theta[[index_of_arm]]$parameter_name}
}
}

\describe{
\item{\code{get_action(context)}}{
here, a policy decides which arm to choose, based on the current values
of its parameters and, potentially, the current context.
}
}

\describe{
\item{\code{set_reward(reward, context)}}{
in \code{set_reward(reward, context)}, a policy updates its parameter values
based on the reward received, and, potentially, the current context.
}
}
}

\references{
Thompson, W. R. (1933). On the likelihood that one unknown probability exceeds another in view of the evidence of two samples. Biometrika, 25(3/4), 285-294.
}
\seealso{
Core contextual classes: \code{\link{Bandit}}, \code{\link{Policy}}, \code{\link{Simulator}},
\code{\link{Agent}}, \code{\link{History}}, \code{\link{Plot}}

Other contextual subclasses: \code{\link{BasicBandit}},
  \code{\link{ContextualBandit}},
  \code{\link{ContextualThompsonSamplingPolicy}},
  \code{\link{ContinuumBandit}},
  \code{\link{DoublyRobustOfflineBandit}},
  \code{\link{EpsilonFirstPolicy}},
  \code{\link{EpsilonGreedyPolicy}},
  \code{\link{Exp3Policy}},
  \code{\link{GittinsBrezziLaiPolicy}},
  \code{\link{LiSamplingOfflineBandit}},
  \code{\link{LifPolicy}},
  \code{\link{LinUCBDisjointPolicy}},
  \code{\link{LinUCBDisjointSmPolicy}},
  \code{\link{LinUCBHybridPolicy}},
  \code{\link{LinUCBHybridSmPolicy}},
  \code{\link{OraclePolicy}}, \code{\link{RandomPolicy}},
  \code{\link{SoftmaxPolicy}},
  \code{\link{SyntheticBandit}},
  \code{\link{ThompsonSamplingPolicy}},
  \code{\link{UCB1Policy}}
}
