% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_cmab_lin_epoch_greedy.R
\name{ContextualEpochGreedyPolicy}
\alias{ContextualEpochGreedyPolicy}
\title{Policy: A Time and Space Efficient Algorithm for Contextual Linear Bandits}
\description{
Policy: A Time and Space Efficient Algorithm for Contextual Linear Bandits
}
\section{Usage}{

\preformatted{
 policy <- ContextualEpochGreedyPolicy$new(sZl = 10)
}
}

\seealso{
Core contextual classes: \code{\link{Bandit}}, \code{\link{Policy}}, \code{\link{Simulator}},
\code{\link{Agent}}, \code{\link{History}}, \code{\link{Plot}}

Bandit subclass examples: \code{\link{BasicBernoulliBandit}}, \code{\link{ContextualLogitBandit}},
\code{\link{OfflineReplayEvaluatorBandit}}

Policy subclass examples: \code{\link{EpsilonGreedyPolicy}}, \code{\link{ContextualLinTSPolicy}}
}
