% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_cmab_epsilon_greedy.R
\name{ContextualEpsilonGreedyPolicy}
\alias{ContextualEpsilonGreedyPolicy}
\title{Policy: Contextual Epsilon Greedy}
\description{
Policy: Contextual Epsilon Greedy
}
\section{Usage}{

\preformatted{
policy <- EpsilonGreedyPolicy(epsilon = 0.1)
}
}

\seealso{
Other contextual subclasses: \code{\link{BasicBandit}},
  \code{\link{ContextualBandit}},
  \code{\link{ContextualDisjointThompsonSamplingPolicy}},
  \code{\link{ContextualThompsonSamplingPolicy}},
  \code{\link{ContinuumBandit}},
  \code{\link{DoublyRobustOfflineBandit}},
  \code{\link{EpsilonFirstPolicy}},
  \code{\link{EpsilonGreedyPolicy}},
  \code{\link{Exp3Policy}},
  \code{\link{GittinsBrezziLaiPolicy}},
  \code{\link{LiSamplingOfflineBandit}},
  \code{\link{LifPolicy}},
  \code{\link{LinUCBDisjointPolicy}},
  \code{\link{LinUCBDisjointSmPolicy}},
  \code{\link{LinUCBHybridPolicy}},
  \code{\link{LinUCBHybridSmPolicy}},
  \code{\link{OraclePolicy}}, \code{\link{RandomPolicy}},
  \code{\link{SoftmaxPolicy}},
  \code{\link{SyntheticBandit}},
  \code{\link{ThompsonSamplingPolicy}},
  \code{\link{UCB1Policy}}
}
