% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandit_cmab_hybrid.R
\name{ContextualHybridBandit}
\alias{ContextualHybridBandit}
\title{Bandit: ContextualHybridBandit}
\description{
Extension of \code{ContextualLogitBandit} modeling hybrid rewards with a combination of unique (or
"disjoint") and shared contextual features.
}
\section{Usage}{

\preformatted{
  bandit <- ContextualHybridBandit$new(k, d, intercept = TRUE)
}
}

\section{Arguments}{


\describe{

\item{\code{k}}{
integer; number of bandit arms
}
\item{\code{d}}{
integer; number of contextual features
}
\item{\code{intercept}}{
logical; if TRUE (default) it adds a constant (1.0) dimension to each context X at the end.
}

}
}

\section{Methods}{


\describe{

\item{\code{new(k, d, intercept = NULL)}}{
generates and instantializes a new \code{Bandit} instance.
For arguments, see Argument section above.
}

\item{\code{get_context(t)}}{
argument:
\itemize{
\item \code{t}: integer, time step \code{t}.
}
returns a named \code{list}
containing the current \code{d x k} dimensional matrix \code{context$X},
the number of arms \code{context$k} and the number of features \code{context$d}.
}

\item{\code{get_reward(t, context, action)}}{
arguments:
\itemize{
\item \code{t}: integer, time step \code{t}.
\item \code{context}: list, containing the current \code{context$X} (d x k context matrix),
\code{context$k} (number of arms) and \code{context$d} (number of context feaures)
(as set by \code{bandit}).
\item \code{action}:  list, containing \code{action$choice} (as set by \code{policy}).
}
returns a named \code{list} containing \code{reward$reward}
}

\item{\code{post_initialization()}}{
initialzes \code{d x k} beta matrix.
}
}
}

\examples{
\donttest{
horizon       <- 800L
simulations   <- 100L



bandit        <- ContextualHybridBandit$new(k = 100, shared_features = 10, unique_features = 2)

#bandit        <- ContextualLogitBandit$new(k = 5, d = 5, intercept = TRUE)

agents        <- list(Agent$new(ContextualThompsonSamplingPolicy$new(delta=0.5, R=0.01, epsilon=0.5), bandit),
                      Agent$new(EpsilonGreedyPolicy$new(0.1), bandit),
                      Agent$new(LinUCBGeneralPolicy$new(0.6), bandit),
                      Agent$new(ContextualEpochGreedyPolicy$new(8), bandit),
                      Agent$new(LinUCBHybridOptimizedPolicy$new(0.6), bandit),
                      Agent$new(LinUCBDisjointOptimizedPolicy$new(0.6), bandit))

simulation     <- Simulator$new(agents, horizon, simulations)
history        <- simulation$run()

plot(history, type = "cumulative", regret = FALSE, rate = TRUE, legend_position = "bottomright")
}
}
\seealso{
Core contextual classes: \code{\link{Bandit}}, \code{\link{Policy}}, \code{\link{Simulator}},
\code{\link{Agent}}, \code{\link{History}}, \code{\link{Plot}}

Bandit subclass examples: \code{\link{BasicBernoulliBandit}}, \code{\link{ContextualLogitBandit}},  \code{\link{OfflinePolicyEvaluatorBandit}}

Policy subclass examples: \code{\link{EpsilonGreedyPolicy}}, \code{\link{ContextualThompsonSamplingPolicy}}
}
