% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_cmab_logit_ts_bootstrap.R
\name{ContextualLogitBTSPolicy}
\alias{ContextualLogitBTSPolicy}
\title{Policy: ContextualLogitBTSPolicy}
\description{
Policy: ContextualLogitBTSPolicy
}
\section{Usage}{

\preformatted{
policy <- ContextualLogitBTSPolicy()
}
}

\seealso{
Core contextual classes: \code{\link{Bandit}}, \code{\link{Policy}}, \code{\link{Simulator}},
\code{\link{Agent}}, \code{\link{History}}, \code{\link{Plot}}

Bandit subclass examples: \code{\link{BasicBernoulliBandit}}, \code{\link{ContextualLogitBandit}},
\code{\link{OfflineReplayEvaluatorBandit}}

Policy subclass examples: \code{\link{EpsilonGreedyPolicy}}, \code{\link{ContextualLinTSPolicy}}
}
