% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_cmab_ts_lin_gaussian.R
\name{ContextualThompsonHybrid}
\alias{ContextualThompsonHybrid}
\title{Policy: A Time and Space Efficient Algorithm for Contextual Linear Bandits}
\description{
Policy: A Time and Space Efficient Algorithm for Contextual Linear Bandits
}
\seealso{
Other contextual subclasses: \code{\link{ContextualEpochGreedyPolicy}},
  \code{\link{ContextualThompsonSamplingPolicy}},
  \code{\link{ContinuumBandit}},
  \code{\link{EpsilonFirstPolicy}},
  \code{\link{EpsilonGreedyPolicy}},
  \code{\link{Exp3Policy}},
  \code{\link{GittinsBrezziLaiPolicy}},
  \code{\link{LiSamplingOfflineBandit}},
  \code{\link{LifPolicy}},
  \code{\link{LinUCBDisjointOptimizedPolicy}},
  \code{\link{LinUCBDisjointPolicy}},
  \code{\link{LinUCBGeneralPolicy}},
  \code{\link{LinUCBHybridOptimizedPolicy}},
  \code{\link{LinUCBHybridPolicy}},
  \code{\link{LinUCBOFULPolicy}},
  \code{\link{OraclePolicy}}, \code{\link{RandomPolicy}},
  \code{\link{SoftmaxPolicy}},
  \code{\link{ThompsonSamplingPolicy}},
  \code{\link{UCB1Policy}}
}
