% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_cmab_ts_probit.R
\name{ContextualTSProbitPolicy}
\alias{ContextualTSProbitPolicy}
\title{Policy: ContextualTSProbitPolicy}
\description{
Makes use of BOPR, ergo only use binary indepependent variables.
}
\section{Usage}{

\preformatted{
policy <- ContextualTSProbitPolicy()
}
}

\seealso{
Core contextual classes: \code{\link{Bandit}}, \code{\link{Policy}}, \code{\link{Simulator}},
\code{\link{Agent}}, \code{\link{History}}, \code{\link{Plot}}

Bandit subclass examples: \code{\link{BasicBernoulliBandit}}, \code{\link{ContextualLogitBandit}},
\code{\link{OfflineReplayEvaluatorBandit}}

Policy subclass examples: \code{\link{EpsilonGreedyPolicy}}, \code{\link{ContextualLinTSPolicy}}
}
