% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandit_continuum.R
\name{ContinuumBandit}
\alias{ContinuumBandit}
\title{Bandit: Continuum}
\description{
ContinuumBandit intro
}
\section{Usage}{

\preformatted{b <- ContinuumBandit$new()

b$reset()

print(b)
}
}

\section{Arguments}{

\describe{
\item{b}{A \code{ContinuumBandit} object.}
}
}

\section{Details}{

\code{$new()} starts a new ContinuumBandit, it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.
}

\examples{
\dontrun{}

}
