% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cx_debug.r
\name{cx_debug}
\alias{cx_debug}
\title{Show debug message}
\usage{
cx_debug(message, ..., debug = .contextual_env$debug, .envir = parent.frame())
}
\arguments{
\item{message}{Debug message}

\item{...}{Additional message components}

\item{debug}{Logical, whether to show debug messages (default: uses global setting)}

\item{.envir}{Environment for string interpolation}
}
\description{
Show debug message
}
