% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cx_inform.r
\name{cx_inform}
\alias{cx_inform}
\title{Show formatted information message}
\usage{
cx_inform(message, ..., verbose = TRUE, .envir = parent.frame())
}
\arguments{
\item{message}{Information message text}

\item{...}{Additional arguments passed to cli functions}

\item{verbose}{Whether to show the information message (TRUE) or suppress it (FALSE)}

\item{.envir}{Environment for glue evaluation}
}
\description{
Show formatted information message
}
