% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cx_title.r
\name{cx_reset_numbering}
\alias{cx_reset_numbering}
\title{Reset section numbering}
\usage{
cx_reset_numbering()
}
\description{
Resets the section numbering system used by cx_title.
Call this at the beginning of a new analysis or report section.
}
