% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{data_table_factors_to_numeric}
\alias{data_table_factors_to_numeric}
\title{Convert all factor columns in data.table to numeric}
\usage{
data_table_factors_to_numeric(dt)
}
\arguments{
\item{dt}{a data.table}
}
\value{
the data.table with column factors converted to numeric
}
\description{
Convert all factor columns in data.table to numeric
}
