% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{dec<-}
\alias{dec<-}
\title{Decrement}
\usage{
dec(x) <- value
}
\arguments{
\item{x}{object to be decremented}

\item{value}{value by which x will be modified}
}
\description{
\code{dec<-} decrements \code{x} by value. Equivalent to \code{x <- x - value.}
}
\examples{
x <- 6:10
dec(x) <- 5
x

}
