% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandit_offline_doubly_robust.R
\name{DoublyRobustOfflineBandit}
\alias{DoublyRobustOfflineBandit}
\title{Bandit: Doubly Robust Offline Bandit}
\description{
\code{DoublyRobustOfflineBandit} evaluates rewards based on offline data
collected by some other policy.
}
\section{Usage}{

\preformatted{
bandit <- DoublyRobustOfflineBandit(data_file, k, d)
}
}

\examples{

horizon            <- 100L

}
\references{
Dudík, M., Langford, J., & Li, L. (2011). Doubly robust policy evaluation and learning. arXiv preprint arXiv:1103.4601.

Dudík, M., Erhan, D., Langford, J., & Li, L. (2014). Doubly robust policy evaluation and optimization. Statistical Science, 485-511.
}
\seealso{
Core contextual classes: \code{\link{Simulator}},
\code{\link{Agent}}, \code{\link{History}}, \code{\link{Plot}}, \code{\link{Policy}}

Other contextual bandits: \code{\link{Bandit}},
  \code{\link{LiSamplingOfflineBandit}}
}
