% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_epsilonfirst.R
\name{EpsilonFirstPolicy}
\alias{EpsilonFirstPolicy}
\title{External EpsilonFirstPolicy}
\description{
EpsilonFirstPolicy intro
}
\section{Usage}{

\preformatted{b <- EpsilonFirstPolicy$new()

b$reset()

print(b)
}
}

\section{Arguments}{

\describe{
\item{b}{A \code{EpsilonFirstPolicy} object.}
}
}

\section{Details}{

\code{$new()} starts a new EpsilonFirstPolicy, it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.
}

\examples{
\dontrun{}

}
