% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.epsilongreedy.R
\name{EpsilonGreedyPolicy}
\alias{EpsilonGreedyPolicy}
\title{External EpsilonGreedyPolicy}
\description{
EpsilonGreedyPolicy intro
}
\section{Usage}{

\preformatted{b <- EpsilonGreedyPolicy$new()

b$reset()

print(b)
}
}

\section{Arguments}{

\describe{
  \item{b}{A \code{EpsilonGreedyPolicy} object.}
}
}

\section{Details}{

\code{$new()} starts a new EpsilonGreedyPolicy, it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.
}

\examples{
\dontrun{}

}
