% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_exp3.R
\name{Exp3Agent}
\alias{Exp3Agent}
\title{External Exp3Agent}
\description{
Exp3Agent intro
}
\section{Usage}{

\preformatted{b <- Exp3Agent$new()

b$reset()

print(b)
}
}

\section{Arguments}{

\describe{
  \item{b}{A \code{Exp3Agent} object.}
}
}

\section{Details}{

\code{$new()} starts a new Exp3Agent, it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.
}

\examples{
\dontrun{}

}
