% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.exp3.R
\name{Exp3Policy}
\alias{Exp3Policy}
\title{External Exp3Policy}
\description{
Exp3Policy intro
}
\section{Usage}{

\preformatted{b <- Exp3Policy$new()

b$reset()

print(b)
}
}

\section{Arguments}{

\describe{
  \item{b}{A \code{Exp3Policy} object.}
}
}

\section{Details}{

\code{$new()} starts a new Exp3Policy, it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.
}

\examples{
\dontrun{}

}
