% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{formatted_difftime}
\alias{formatted_difftime}
\title{Format difftime objects}
\usage{
formatted_difftime(x)
}
\arguments{
\item{x}{difftime object}
}
\value{
string "days, h:mm:ss.ms"
}
\description{
Format difftime objects
}
