% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gammaAtv.R
\name{gammaAtv}
\alias{gammaAtv}
\alias{gammam}
\alias{extend.gammam}
\alias{extend.expAtv}
\title{Evaluate Transition Matrix After a Gamma-Distributed Time}
\usage{
gammaAtv(A, scale, shape, v, tol = 1e-06, verbose = FALSE)

gammam(A, scale, shape, ...)

extend.gammam(P, a, da, eps = da/(a + da), tol = 1e-08)

extend.expAtv(A, scale, dscale, gAtv, tol = 1e-08, verbose = FALSE, shape)
}
\arguments{
\item{scale}{Scale parameter (=1/beta).}

\item{shape}{Shape parameter for Gamma (=alpha).}

\item{scale.t}{Overall rate of the transitions (a time scaling).

Note that diag(A) is not used.}
}
\description{
Evaluate transition matrix after gamma-distributed time, multiplied by the
matrix `v`.
}
\details{
The math behind this: let X be a Markov chain with generator matrix A
 and T ~ Gamma(shape=alpha,rate=beta)
 so that
   P(X_T=y|X_0=x) 
       = \int_0^\infty ( \beta^\alpha / \Gamma(\alpha) ) t^{\alpha-1} \exp( - \beta t + A t ) dt
 which if (\lambda,v) is the eigendecomposition of A is
                  = sum_i v_i v_i^T \beta^\alpha / (\beta - \lambda_i)^\alpha

Note that if N(t) ~ Pois(lambda t)
 and T ~ Gamma(shape=alpha,rate=beta)
 then N(T) ~ NegativeBinomial(alpha,lambda/(lambda+beta))
 i.e. P(N(T)=n) = Gamma(n+alpha)/(Gamma(alpha)*factorial(n)) 
                  * (beta/(beta+lambda))^alpha * (lambda/(beta+lambda))^n
 and that
     P(N(T)=n) = (lambda * (n+alpha-1)) / ( n * (beta+lambda) ) * P(N(T)=n-1) .

Furthermore, 
 if  T ~ Exponential(rate=beta),
 we've already sampled N(T),
 and want to get a sample of N(S), 
 where S ~ Exponential(rate=beta+u)
 then we can let
   S = T + sum_{k=0}^M T_k
 where M is Geometric(beta/(beta+u)) and T_k are independent copies of T.
This implies that we can write
   N(S) = N(T) + N' 
 where N' ~ NegBinom(M,lambda/(lambda+beta))
 which is the same as
       N' ~ NegBinom(M,lambda/(lambda+beta+u))
}
\section{Functions}{
\itemize{
\item \code{gammam}: Return the whole transition matrix

\item \code{extend.gammam}: Extend a transition matrix from gammam() to a longer time.

\item \code{extend.expAtv}: Extend a computed gammaAtv to a longer time.
}}

