% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandit_gaussian.R
\name{GaussianBandit}
\alias{GaussianBandit}
\title{GaussianBandit}
\description{
Bandit of type: Basic Gaussian
}
\section{Usage}{

\preformatted{p <- process$new(command = NULL, args, commandline = NULL,
                 stdout = TRUE, stderr = TRUE)
}
}

\section{Arguments}{

\describe{
  \item{k}{An integer: how many arms.}
}
}

\section{Details}{

\code{$new()} starts a new process, it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.

\code{$reset()} Resets the bandit.

\code{$pull(action)} Returns reward and True if action is optimal.
}

\examples{

}
