% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genmatrix.R
\docType{class}
\name{genmatrix-class}
\alias{genmatrix-class}
\title{Generator Matrix Class (genmatrix)}
\description{
genmatrix extends the sparse matrix class, carrying along more information.
}
\details{
genmatrix gives the instantaneous rate for going from patterns x -> y
It is laid out in classical Markov fashion, with the rows indexing the "from" states
(headpats) and columns indexing the "to" states (tailpats)
}
