% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/context_classes.R
\docType{methods}
\name{longwin,genmatrix-method}
\alias{longwin,genmatrix-method}
\alias{nmuts,genmatrix-method}
\alias{nsel,genmatrix-method}
\alias{fixparams,genmatrix-method}
\title{Methods for genmatix objects}
\usage{
\S4method{longwin}{genmatrix}(x)

\S4method{nmuts}{genmatrix}(x)

\S4method{nsel}{genmatrix}(x)

\S4method{fixparams}{genmatrix}(x)
}
\description{
Methods for genmatix objects
}
