% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{get.root.distrn}
\alias{get.root.distrn}
\title{Product Distribution for the Root of a Tree}
\usage{
get.root.distrn(initfreqs, initfreq.index)
}
\description{
Given initial frequencies find the distribution over longpats
(the product measure) #' with helper matrix computed in product.index()
}
