% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{get_arm_context}
\alias{get_arm_context}
\title{Return context vector of an arm}
\usage{
get_arm_context(context, arm, select_features = NULL)
}
\arguments{
\item{context}{a context list containing a d x k Matrix or
d dimensional context vector X, the number of features d and
number of arms k.}

\item{arm}{index of arm.}

\item{select_features}{indices of to be returned features.}
}
\value{
Vector that represents context related to an arm
}
\description{
Given d x k matrix or d dimensional vector X,
returns a vector with arm's context.
}
