% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{get_global_seed}
\alias{get_global_seed}
\title{Lookup .Random.seed in global environment}
\usage{
get_global_seed()
}
\value{
an integer vector, containing the random number generator (RNG) state for random number generation
}
\description{
Lookup .Random.seed in global environment
}
