% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making_matrices.R
\name{getmutmats}
\alias{getmutmats}
\title{Set Up a List of Mutation Matrices}
\usage{
getmutmats(mutpats, patterns, boundary = c("none", "wrap"),
  do.parallel = ("parallel" \%in\% .packages(all.available = TRUE)),
  numcores = if (do.parallel) {     parallel::detectCores() } else {     1 })
}
\description{
The purpose of this function is to construct a translation table between
mutpats and the entries of the generator matrix that can happen as a
consequence of those mutpats.
Specifically, returns i and j's indexing the generator matrix in terms of the given patterns.
that is, given a list of mutation patterns,
  which can be either pairs or lists of pairs,
return a corresponding list of two-column matrices with (1-based) indices of changes corresponding to mutation patterns
  i.e. if (i,j) is a row of output[[k]], then patterns[j] can be obtained from patterns[i]
  by performing a substitution from mutpats[[k]] at some location within the string.
}
