% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{getmutpats}
\alias{getmutpats}
\title{Get all k-mer changes}
\usage{
getmutpats(bases, patlen, nchanges = 1)
}
\description{
Finds all kmer -> kmer changes where k <= `patlen`
that involve at most nchanges changes.
}
