% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/making_matrices.R
\name{getselmatches}
\alias{getselmatches}
\title{Find Matches for Selection Patterns}
\usage{
getselmatches(selpats, patterns, selfactors, boundary = c("none", "wrap"),
  names = FALSE, do.parallel = ("parallel" \%in\% .packages(all.available =
  TRUE)), numcores = if (do.parallel) {     parallel::detectCores() } else {  
    1 })
}
\description{
each element gets one parameter
selmatches[i,j] is the sum of selfactors[[i]] multiplied by the number of times the corresponding thing in selpat[[i]] matches pattern[j]
 ... so if selfactors is all 1's, then :
    selmatches[i,j] is number of times anything in selpat[[i]] matches pattern[j]
}
