% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_matrix.R
\name{getupdowntrans}
\alias{getupdowntrans}
\title{Finds the matrix of conditional probabilities of y given x across a two-branch tree.}
\usage{
getupdowntrans(genmatrix, projmatrix, mutrates, selcoef, initfreqs,
  tlens = c(1, 1), ...)
}
\value{
Matrix with entry [x,y] the probability of seeing y on the "down
  branch" given x was seen on the "up" branch.
}
\description{
arguments are lists of two: first the "up" branch (leading from simpler summaries), second the "down"
}
