% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_mab_gittins_bl.R
\name{GittinsBrezziLaiPolicy}
\alias{GittinsBrezziLaiPolicy}
\title{Policy: Gittins Approximation algorithm for choosing arms in a MAB problem.}
\description{
\code{GittinsBrezziLaiPolicy} Algorithm based on Brezzi and Lai (2002) "Optimal learning
and experimentation in bandit problems."
The algorithm provides an approximation of the Gittins index, by specifying
a closed-form expression, which is a function of the discount factor, and
the number of successes and failures associated with each arm.
}
\seealso{
Core contextual classes: \code{\link{Bandit}}, \code{\link{Policy}}, \code{\link{Simulator}},
\code{\link{Agent}}, \code{\link{History}}, \code{\link{Plot}}

Bandit subclass examples: \code{\link{BasicBernoulliBandit}}, \code{\link{ContextualLogitBandit}},  \code{\link{OfflinePolicyEvaluatorBandit}}

Policy subclass examples: \code{\link{EpsilonGreedyPolicy}}, \code{\link{ContextualThompsonSamplingPolicy}}
}
