% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{inc<-}
\alias{inc<-}
\title{Increment}
\usage{
inc(x) <- value
}
\arguments{
\item{x}{object to be incremented}

\item{value}{value by which x will be modified}
}
\description{
\code{inc<-} increments \code{x} by value. Equivalent to \code{x <- x + value.}
}
\examples{
x <- 1:5
inc(x) <- 5
x

}
