% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{ind}
\alias{ind}
\title{On-the-fly indicator function for use in formulae}
\usage{
ind(cond)
}
\arguments{
\item{cond}{a logical condition to be evaluated}
}
\value{
a binary (0/1) coded variable indicating whether the condition is true
}
\description{
On-the-fly indicator function for use in formulae
}
