% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{inv}
\alias{inv}
\title{Inverse from Choleski (or QR) Decomposition.}
\usage{
inv(M)
}
\arguments{
\item{M}{matrix}
}
\description{
Invert a symmetric, positive definite square matrix from its Choleski decomposition.
}
\examples{
inv(cbind(1, 1:3, c(1,3,7)))

}
