% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{inv_logit}
\alias{inv_logit}
\title{Inverse Logit Function}
\usage{
inv_logit(x)
}
\arguments{
\item{x}{A numeric object.}
}
\value{
An object of the same type as x containing the inverse logits of the input values.
}
\description{
Given a numeric object return the inverse logit of the values.
}
