% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_lif.R
\name{LifPolicy}
\alias{LifPolicy}
\title{Policy: LiF}
\description{
LifPolicy intro
}
\section{Usage}{

\preformatted{b <- LifPolicy$new()

b$reset()

print(b)
}
}

\section{Arguments}{

\describe{
\item{b}{A \code{LifPolicy} object.}
}
}

\section{Details}{

\code{$new()} starts a new LifPolicy, it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.
}

\examples{
\dontrun{}

}
