% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihood.R
\name{likelihood.function}
\alias{likelihood.function}
\title{Make a Likelihood Function}
\usage{
likelihood.function(longwin, shortwin, leftwin, config, seq)
}
\arguments{
\item{longwin}{The length of the longest window used in Tmers.}

\item{shortwin}{The length of the short window used in Tmers.}

\item{leftwin}{The offset of short window used.}

\item{config}{A list of arguments to `makegenmatrix()` except patlen.}

\item{seq}{The sequence to compute the likelihood of.}
}
\value{
A function.
}
\description{
Returns a log likelihood function appropriate for optimization,
which is a function of a single vector, concatenated, of:
  mutrates, selcoef, fixparams
}
\details{
The likelihood is the ratio of likelihoods with (long,short) 
and (long,short-1) window lengths.

The length of time is fixed to be 1.

Note this returns log-likelihood, *not* negative log-likelihood.
}
\examples{

config <- list(mutpats=list(list(c("XX","OO")),list(c("X","O"),c("O","X"))), selpats=list(), bases=c("X","O"), fixfn=null.fixfn, selfactors=list())
seq <- list( 
  initseq=
     "OOXXXXOOXXOXOXOXOOXOOOXOXOXXOOXXOOOOOOXOXOXXXXXXXOXOXOOXXXXOXOXXOOOXXXXOXOOXOXOOOXOXOXOOXXXOOOXOXOXXXOOOOXXXXOOOXOOOOXOOOXXOXXOXOOXXOOXXXOXOXOXXOOOXOOOOXOOXXXXOXXXXOOXOOOOXOOXXOXOXXOOXXXOXOXXXOOOOOOOX",
  finalseq=
     "OOXXOXOXXOXOXXOXOOOXOOXOOXXOOOXXXOOOOOOOXXXXXXXXOOXXOXOXXXOOXXXOXOOXOXXXXOOXOOXOOOXXOOXOXOXOXOXOXOXXXOOOOXXXOOXOOOOOXOXOOXXOXXOXOOXOXOXXOXXOOXXXOOOXOOOOOXOXXXXOXXXXOOOOXOOXOOXXOXOXXOOXXXOOXXXOXOOOOOOX")
lfn <- likelihood.function(5,3,1,config,seq)
# arguments to lfn are: the XX -> OO mutation rate, and the X <-> O mutation rate
with(environment(lfn), mutnames(genmatrix@mutpats))
lfn(c(1.5,2.0))

}
