% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandit_log.R
\name{LiLogBandit}
\alias{LiLogBandit}
\title{External LiLogBandit}
\description{
LiLogBandit intro
}
\section{Usage}{

\preformatted{b <- LiLogBandit$new()

b$reset()

print(b)
}
}

\section{Arguments}{

\describe{
  \item{b}{A \code{LiLogBandit} object.}
}
}

\section{Details}{

\code{$new()} starts a new LiLogBandit, it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.
}

\examples{
\dontrun{}

}
