% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.linucb.R
\name{LinUCBAgent}
\alias{LinUCBAgent}
\title{External LinUCBAgent}
\description{
LinUCBAgent intro
}
\section{Usage}{

\preformatted{b <- LinUCBAgent$new()

b$reset()

print(b)
}
}

\section{Arguments}{

\describe{
  \item{b}{A \code{LinUCBAgent} object.}
}
}

\section{Details}{

\code{$new()} starts a new LinUCBAgent, it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.
}

\examples{
\dontrun{}

}
