% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_linucb.R
\name{LinUCBPolicy}
\alias{LinUCBPolicy}
\title{External LinUCBPolicy}
\description{
LinUCBPolicy intro
}
\section{Usage}{

\preformatted{b <- LinUCBPolicy$new()

b$reset()

print(b)
}
}

\section{Arguments}{

\describe{
\item{b}{A \code{LinUCBPolicy} object.}
}
}

\section{Details}{

\code{$new()} starts a new LinUCBPolicy, it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.
}

\examples{
\dontrun{}

}
