% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genmatrix.R
\name{meangenmatrix}
\alias{meangenmatrix}
\title{Create an Boundary-Averaged Generator Matrix}
\usage{
meangenmatrix(leftwin, rightwin, patlen, ...)
}
\description{
Create a generator matrix that averages over possible adjacent states
}
