% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandit_cmab_mixed.R
\name{MixedContextualBandit}
\alias{MixedContextualBandit}
\title{Bandit: MixedContextualBandit}
\description{
MixedContextualBandit intro
}
\details{
Following Allen Day DataGenerator
}
\section{Usage}{

\preformatted{b <- MixedContextualBandit$new()

b$reset()

print(b)
}
}

\section{Arguments}{

\describe{
\item{b}{A \code{MixedContextualBandit} object.}
}
}

\section{Details}{

\code{$new()} starts a new MixedContextualBandit, it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.
}

\seealso{
Core context_commonual classes: \code{\link{Bandit}}, \code{\link{Policy}}, \code{\link{Simulator}},
\code{\link{Agent}}, \code{\link{History}}, \code{\link{Plot}}

Other context_commonual subclasses: \code{\link{BasicContextualBandit}}
}
