% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{mutnames}
\alias{mutnames}
\title{Print a Mutation Pattern List in Readable Form}
\usage{
mutnames(mutpats, pretty = TRUE)
}
\description{
Stringify a list of mutpat lists,
 if pretty=TRUE, then using the names if they are available
}
