% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{mutpatchanges}
\alias{mutpatchanges}
\title{Find all possible mutational changes}
\usage{
mutpatchanges(mutpats)
}
\description{
return list of single-base changes for each list of patterns in mutpats
a `mutpat` is a 2-element string list of the form (from, to)
}
