% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mv_title.r
\name{mv_reset_numbering}
\alias{mv_reset_numbering}
\title{Reset section numbering}
\usage{
mv_reset_numbering()
}
\description{
Resets the section numbering system used by mv_title.
Call this at the beginning of a new analysis or report section.
}
