% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{mvrnorm}
\alias{mvrnorm}
\title{Simulate from a Multivariate Normal Distribution}
\usage{
mvrnorm(n, mu, sigma)

mvrnorm(n, mu, sigma)
}
\arguments{
\item{n}{the number of samples required.}

\item{mu}{a vector giving the means of the variables.}

\item{sigma}{a positive-definite symmetric matrix specifying the covariance matrix of the variables.}

\item{n}{the number of samples required.}

\item{mu}{a vector giving the means of the variables.}

\item{sigma}{a positive-definite symmetric matrix specifying the covariance matrix of the variables.}
}
\value{
If \code{n = 1} a vector of the same length as \code{mu}, otherwise an \code{n} by
\code{length(mu)} matrix with one sample in each row.

If \code{n = 1} a vector of the same length as \code{mu}, otherwise an \code{n} by
\code{length(mu)} matrix with one sample in each row.
}
\description{
Produces one or more samples from the specified
multivariate normal distribution.

Produces one or more samples from the specified
multivariate normal distribution.
}
