% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-output.R
\name{nodenames}
\alias{nodenames}
\alias{nodenames<-}
\alias{rootname}
\alias{rootname<-}
\alias{edge.labels}
\alias{get.root}
\alias{get.parent}
\alias{get.cherries}
\title{Show the names of the nodes of a tree.}
\usage{
nodenames(tr)

nodenames(tr) <- value

rootname(tr)

rootname(tr) <- value

edge.labels(tr)

get.root(tr)

get.parent(node, tr)

get.cherries(node, tr)
}
\description{
Show the names of the nodes of a tree.
}
\section{Functions}{
\itemize{
\item \code{nodenames<-}: Assign node names to a tree.

\item \code{rootname}: Find the name of a root of a tree.

\item \code{rootname<-}: Assign the name of a root of a tree.

\item \code{edge.labels}: Return the name of the downstream node of each edge.

\item \code{get.root}: Return the index of the root in (tips,nodes) order

\item \code{get.parent}: Return index of the parent of node in (tips,nodes) order

\item \code{get.cherries}: Return pairs in node (or NA if none exists)
}}

