% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{ones_in_zeroes}
\alias{ones_in_zeroes}
\title{A vector of zeroes and ones}
\usage{
ones_in_zeroes(vector_length, index_of_one)
}
\arguments{
\item{vector_length}{How long will the vector be?}

\item{index_of_one}{Where to insert the one?}
}
\value{
Vector of zeroes with one(s) at given index position(s)
}
\description{
A vector of zeroes and ones
}
