% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_oracle.R
\name{OraclePolicy}
\alias{OraclePolicy}
\title{External OraclePolicy}
\description{
OraclePolicy intro
}
\section{Usage}{

\preformatted{b <- OraclePolicy$new()

b$reset()

print(b)
}
}

\section{Arguments}{

\describe{
  \item{b}{A \code{OraclePolicy} object.}
}
}

\section{Details}{

\code{$new()} starts a new OraclePolicy, it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.
}

\examples{
\dontrun{}

}
