% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-output.R
\name{parse.fixfn}
\alias{parse.fixfn}
\title{Get a Fixation Function}
\usage{
parse.fixfn(fixfn, fixfn.params)
}
\description{
Turn fixfn into an actual function
  either by looking it up as a name
  or parsing it directly
also, check the arguments match fixfn.params.
}
