% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-output.R
\name{.parse.selpats}
\alias{.parse.selpats}
\title{Parse selpats.}
\usage{
.parse.selpats(config)
}
\description{
check if (selpats,selfactors) info is combined into selpats
 and separate out if so
So: selpats should be EITHER:
  - a list of character vectors, OR
  - a list of named lists of numeric values
}
