% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandit_parser.R
\name{ParserBandit}
\alias{ParserBandit}
\title{External ParserBandit}
\description{
ParserBandit intro
}
\section{Usage}{

\preformatted{b <- ParserBandit$new()

b$reset()

print(b)
}
}

\section{Arguments}{

\describe{
  \item{b}{A \code{ParserBandit} object.}
}
}

\section{Details}{

\code{$new()} starts a new ParserBandit, it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.
}

\examples{
\dontrun{}

}
