% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_matrix.R
\name{peel.transmat.compute}
\alias{peel.transmat.compute}
\title{Compute probabilities of all combinations, on a tree}
\usage{
peel.transmat.compute(setup, models, genmatrices, root.distrn, tlens,
  debug = FALSE, return.list = TRUE)
}
\description{
Compute probabilities of all combinations, on a tree
}
