% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_matrix.R
\name{peel.transmat.setup}
\alias{peel.transmat.setup}
\title{Set up for peel.transmat}
\usage{
peel.transmat.setup(tree, rowtaxon, coltaxa, models, genmatrices, projmatrix,
  tlens = tree$edge.length, return.list = FALSE, debug = FALSE)
}
\description{
Figures out what order to do things in.
}
