% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Plot}
\alias{Plot}
\title{External Plot}
\description{
Plot intro
}
\section{Usage}{

\preformatted{b <- Plot$new()

b$reset()

print(b)
}
}

\section{Arguments}{

\describe{
\item{b}{A \code{Plot} object.}
}
}

\section{Details}{

\code{$new()} starts a new Plot, it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.
}

\examples{
\dontrun{}

}
