% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixfns.R
\name{popgen.fixfn}
\alias{popgen.fixfn}
\title{Population Genetics Fixation Function}
\usage{
popgen.fixfn(ds, Ne, ...)
}
\description{
Total influx of fixation given selection coefficient (s[to] - s[from]) difference ds
}
