% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_counts.R
\name{predictcounts.context}
\alias{predictcounts.context}
\title{Predict counts for a 'context' model}
\usage{
predictcounts.context(model, longwin = NULL, shortwin = NULL,
  leftwin = NULL, initcounts = rowSums(model), mutrates = model@mutrates,
  selcoef = model@selcoef, genmatrix = model@genmatrix,
  projmatrix = model@projmatrix, params = model@params, tlen = 1)
}
\description{
Predict counts for a 'context' model
}
