% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_counts.R
\name{predicttreecounts}
\alias{predicttreecounts}
\title{Compute expected counts of paired patterns}
\usage{
predicttreecounts(shortwin, leftwin = 0, rightwin = 0, initcounts, mutrates,
  selcoef, tlens, genmatrix, projmatrix, initfreqs, patcomp, ...)
}
\description{
Compute expected counts of paired patterns
}
