% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_random.R
\name{RandomPolicy}
\alias{RandomPolicy}
\title{External RandomPolicy}
\description{
RandomPolicy intro
}
\section{Usage}{

\preformatted{b <- RandomPolicy$new()

b$reset()

print(b)
}
}

\section{Arguments}{

\describe{
  \item{b}{A \code{RandomPolicy} object.}
}
}

\section{Details}{

\code{$new()} starts a new RandomPolicy, it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.
}

\examples{
\dontrun{}

}
