% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-output.R
\name{read.config.counts}
\alias{read.config.counts}
\title{Read in the header configuration line(s) in a counts file.}
\usage{
read.config.counts(infile)
}
\arguments{
\item{infile}{The file name.}
}
\value{
The parsed JSON header.
}
\description{
Read in the header configuration line(s) in a counts file.
}
