% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-output.R
\name{read.config}
\alias{read.config}
\title{Read in JSON Config File}
\usage{
read.config(configfile, quiet = FALSE, json)
}
\value{
A model configuration (list).
}
\description{
... or directly from a text string. Automatically fills in zero-length
parameters, and fills in selfactors if it isn't there, or moves values from
selpats if it is.
}
\examples{

model_json <- '
{
    "bases" : [ "X", "O" ],
    "initfreqs" : [ 0.5, 0.5 ],
    "mutpats" : [
        [ [ "XO", "OX" ] ]
    ],
    "mutrates" : [ 1 ],
    "mutrates.scale" : [ 0.1 ]
}
'
model_config <- read.config(json=model_json)

}
