% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{reorder.counts}
\alias{reorder.counts}
\title{Reorder Counts by Taxon Reordering}
\usage{
\method{reorder}{counts}(counts, new.ord)
}
\description{
Rearrange the columns of counts to a new ordering of taxa
new.ord should be a vector of taxon names, i.e. a reordering of colnames(counts@colpatterns)
}
