% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{resid.context}
\alias{resid.context}
\title{used in setMethod}
\usage{
\method{resid}{context}(object, counts = object@counts,
  genmatrix = object@genmatrix, pretty = FALSE)
}
\description{
used in setMethod
}
