% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{sample_one_of}
\alias{sample_one_of}
\title{Sample one of}
\usage{
sample_one_of(x)
}
\arguments{
\item{x}{A vector of one or more elements from which to choose}
}
\value{
One value, drawn from x.
}
\description{
Sample one of the values in a vector.
}
