% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/patterns.R
\name{selnames}
\alias{selnames}
\title{Print a Selection Pattern List in Readable Form}
\usage{
selnames(selpats, pretty = TRUE)
}
\description{
Stringify a list of selpats,
 but use the names if they are available
}
