% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{set_global_seed}
\alias{set_global_seed}
\title{Set .Random.seed to a pre-saved value}
\usage{
set_global_seed(x)
}
\arguments{
\item{x}{integer vector}
}
\description{
Set .Random.seed to a pre-saved value
}
