% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixfns.R
\name{shape.fixfn}
\alias{shape.fixfn}
\title{Shape Fixation Function}
\usage{
shape.fixfn(ds, Ne, ...)
}
\description{
Total influx of fixation given shape difference ds = (s[to] - s[from]),
  which assuming change is BAD, means a selection differential of (-1)*abs(ds)
}
