% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{sherman_morrison}
\alias{sherman_morrison}
\title{Inverse from Choleski (or QR) Decomposition.}
\usage{
sherman_morrison(M_inv, x)
}
\arguments{
\item{M_inv}{matrix}

\item{x}{column vector}
}
\value{
(M + x*x')^-1 computed using Sherman-Morrison formula
}
\description{
Invert a symmetric, positive definite square matrix from its Choleski decomposition.
}
