% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{sherman_morrisson}
\alias{sherman_morrisson}
\title{Sherman-Morrisson inverse}
\usage{
sherman_morrisson(inv, x)
}
\arguments{
\item{inv}{to be updated inverse matrix}

\item{x}{column vector to update inv with}
}
\description{
Sherman-Morrisson inverse
}
