% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{Simulator}
\alias{Simulator}
\title{External Simulator}
\description{
Simulator intro
}
\section{Usage}{

\preformatted{b <- Simulator$new()

b$reset()

print(b)
}
}

\section{Arguments}{

\describe{
\item{b}{A \code{Simulator} object.}
}
}

\section{Details}{

\code{$new()} starts a new Simulator, it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.
}

\examples{
\dontrun{}

}
