% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator_azure.R
\name{SimulatorAzure}
\alias{SimulatorAzure}
\title{External SimulatorAzure}
\description{
SimulatorAzure intro
}
\section{Usage}{

\preformatted{b <- SimulatorAzure$new()

b$reset()

print(b)
}
}

\section{Arguments}{

\describe{
  \item{b}{A \code{SimulatorAzure} object.}
}
}

\section{Details}{

\code{$new()} starts a new SimulatorAzure, it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.
}

\examples{
\dontrun{}

}
