% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator_basic.R
\name{SimulatorBasic}
\alias{SimulatorBasic}
\title{External SimulatorBasic}
\description{
SimulatorBasic intro
}
\section{Usage}{

\preformatted{b <- SimulatorBasic$new()

b$reset()

print(b)
}
}

\section{Arguments}{

\describe{
  \item{b}{A \code{SimulatorBasic} object.}
}
}

\section{Details}{

\code{$new()} starts a new SimulatorBasic, it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.
}

\examples{
\dontrun{}

}
