% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulator.R
\name{SimulatorParallel}
\alias{SimulatorParallel}
\title{External SimulatorParallel}
\description{
SimulatorParallel intro
}
\section{Usage}{

\preformatted{b <- SimulatorParallel$new()

b$reset()

print(b)
}
}

\section{Arguments}{

\describe{
  \item{b}{A \code{SimulatorParallel} object.}
}
}

\section{Details}{

\code{$new()} starts a new SimulatorParallel, it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.
}

\examples{
\dontrun{}

}
