% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{sum_of}
\alias{sum_of}
\title{Sum of list}
\usage{
sum_of(x)
}
\arguments{
\item{x}{List}
}
\description{
Returns the sum of the values of the elements of a list \code{x}.
}
\examples{

theta = list(par_one = list(1,2,3), par_two = list(2,3,4))
sum_of(theta$par_one)

}
