% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_generic.R
\name{summary.history}
\alias{summary.history}
\alias{summary.History}
\title{Summary Method for Contextual History}
\usage{
\method{summary}{History}(object, ...)
}
\arguments{
\item{object}{A \code{History} object.}

\item{...}{Further summary parameters.}
}
\description{
summary.history, a method for the summary generic. It is designed for a quick summary of History data.
}
\seealso{
Core contextual classes: \code{\link{Bandit}}, \code{\link{Policy}}, \code{\link{Simulator}},
\code{\link{Agent}}, \code{\link{History}}, \code{\link{Plot}}

Bandit subclass examples: \code{\link{BasicBernoulliBandit}}, \code{\link{ContextualLogitBandit}},  \code{\link{OfflineReplayEvaluatorBandit}}

Policy subclass examples: \code{\link{EpsilonGreedyPolicy}}, \code{\link{ContextualThompsonSamplingPolicy}}
}
