% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandit.synthetic.R
\name{SyntheticBandit}
\alias{SyntheticBandit}
\title{External SyntheticBandit}
\description{
SyntheticBandit intro
}
\section{Usage}{

\preformatted{b <- SyntheticBandit$new()

b$reset()

print(b)
}
}

\section{Arguments}{

\describe{
  \item{b}{A \code{SyntheticBandit} object.}
}
}

\section{Details}{

\code{$new()} starts a new SyntheticBandit, it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.
}

\examples{
\dontrun{}

}
