% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_thompsonsampling.R
\name{ThompsonSamplingPolicy}
\alias{ThompsonSamplingPolicy}
\title{External ThompsonSamplingPolicy}
\description{
ThompsonSamplingPolicy intro
}
\section{Usage}{

\preformatted{b <- ThompsonSamplingPolicy$new()

b$reset()

print(b)
}
}

\section{Arguments}{

\describe{
  \item{b}{A \code{ThompsonSamplingPolicy} object.}
}
}

\section{Details}{

\code{$new()} starts a new ThompsonSamplingPolicy, it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.
}

\examples{
\dontrun{}

}
