% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-output.R
\name{treeify.config}
\alias{treeify.config}
\title{Treeify a Configuration.}
\usage{
treeify.config(config, tlen = NULL)
}
\value{
A model configuration (list).
}
\description{
Turn Newick tree representation into a tree, 
  and do associated error checks.
no tree at all =>  stick tree.
}
