% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tuplecounts.R, R/context_classes.R
\docType{methods}
\name{rowtaxon}
\alias{rowtaxon}
\alias{rowtaxon,tuplecounts-method}
\alias{coltaxa}
\alias{coltaxa,tuplecounts-method}
\alias{colpatterns}
\alias{colpatterns,tuplecounts-method}
\alias{counts}
\alias{counts,tuplecounts-method}
\alias{countframe}
\alias{countframe,tuplecounts-method}
\alias{dim,tuplecounts-method}
\alias{dimnames,tuplecounts-method}
\alias{dimnames<-,tuplecounts,ANY-method}
\alias{dimnames<-,tuplecounts-method}
\alias{as.matrix,tuplecounts-method}
\alias{as.vector,tuplecounts-method}
\alias{head,tuplecounts-method}
\alias{image,tuplecounts-method}
\alias{longwin,tuplecounts-method}
\alias{shortwin,tuplecounts-method}
\alias{leftwin,tuplecounts-method}
\alias{rowSums,tuplecounts-method}
\alias{image,tuplecounts-method}
\title{This method repackages the counts in a more friendly-looking data frame}
\usage{
rowtaxon(x)

\S4method{rowtaxon}{tuplecounts}(x)

coltaxa(x)

\S4method{coltaxa}{tuplecounts}(x)

colpatterns(x)

\S4method{colpatterns}{tuplecounts}(x)

counts(x)

\S4method{counts}{tuplecounts}(x)

countframe(x, ...)

\S4method{countframe}{tuplecounts}(x, include.zeros = FALSE)

\S4method{dim}{tuplecounts}(x)

\S4method{dimnames}{tuplecounts}(x)

\S4method{dimnames}{tuplecounts,ANY}(x) <- value

\S4method{as.matrix}{tuplecounts}(x)

\S4method{as.vector}{tuplecounts}(x)

\S4method{head}{tuplecounts}(x)

\S4method{image}{tuplecounts}(x)

\S4method{longwin}{tuplecounts}(x)

\S4method{shortwin}{tuplecounts}(x)

\S4method{leftwin}{tuplecounts}(x)

\S4method{rowSums}{tuplecounts}(x)

\S4method{image}{tuplecounts}(x)
}
\arguments{
\item{x}{A tuplecounts object.}

\item{include.zeros}{Whether to include zero entries in the result.}
}
\description{
This method repackages the counts in a more friendly-looking data frame
}
