% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_matrix.R
\name{upbranch}
\alias{upbranch}
\title{"prune" a dangling edge in the "up" direction --
   return diag(rootfreqs) * e^( tlen * genmatrix(mut,sel) ) %*% tipmatrix .
So, result has
  upbranch[i,j] = rootfreqs[i] * E[ tipmatrix[X(tlen),j] | X(0) = i ]}
\usage{
upbranch(genmatrix, tipmatrix, rootfreqs, mutrates, selcoef, tlen, ...)
}
\description{
"prune" a dangling edge in the "up" direction --
   return diag(rootfreqs) * e^( tlen * genmatrix(mut,sel) ) %*% tipmatrix .
So, result has
  upbranch[i,j] = rootfreqs[i] * E[ tipmatrix[X(tlen),j] | X(0) = i ]
}
