% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genmatrix.R
\name{update}
\alias{update}
\title{Update the Entries of a Generator Matrix With New Rates.}
\usage{
update(G, mutrates, selcoef, ...)
}
\description{
Calculate the new entries of `x` from muttrans et al.
use like: genmatrix@x <- update(genmatrix,...)
}
