% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{var_welford}
\alias{var_welford}
\title{Welford's variance}
\usage{
var_welford(z)
}
\arguments{
\item{z}{vector}
}
\value{
variance
}
\description{
Welford described a method for 'robust' one-pass computation of the
standard deviation. By 'robust', we mean robust to round-off caused
by a large shift in the mean.
}
