% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{which_max_list}
\alias{which_max_list}
\title{Get maximum value in list}
\usage{
which_max_list(x, equal_is_random = TRUE)
}
\arguments{
\item{x}{vector of values}

\item{equal_is_random}{boolean}
}
\description{
Returns the index of the maximum value in list \code{x}.
}
\details{
If there is a tie and \code{equal_is_random} is \code{TRUE},
the index of one of the tied maxima is returned at random.

If \code{equal_is_random} is \code{FALSE},
the maximum with the lowest index number is returned.
}
\examples{

theta = list(par_one = list(1,2,3), par_two = list(2,3,4))
which_max_list(theta$par_one)

}
