% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{which_max_tied}
\alias{which_max_tied}
\title{Get maximum value randomly breaking ties}
\usage{
which_max_tied(x, equal_is_random = TRUE)
}
\arguments{
\item{x}{vector of values}

\item{equal_is_random}{boolean}
}
\description{
Returns the index of the maximum value in vector \code{vec}.
}
\details{
If there is a tie, and equal_is_random is TRUE,
the index of one of the tied maxima is returned at random. Otherwise,
the value with the lowest index is returned.
}
