% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapsubstr.R
\name{wrapsubstr}
\alias{wrapsubstr}
\alias{wrapsubstr<-}
\title{Circular Substring}
\usage{
wrapsubstr(x, start, stop)

wrapsubstr(x, start, stop) <- value
}
\description{
Finds the substring treating x as a circular string
}
\section{Functions}{
\itemize{
\item \code{wrapsubstr<-}: Assign to circular substring
}}

