% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-output.R
\name{write.counts}
\alias{write.counts}
\title{Convert Tuplecounts object to data frame
Write a Tuplecounts object to file}
\usage{
write.counts(x, filename, ...)
}
\arguments{
\item{x}{A tuplecounts object.}

\item{filename}{A filename.}

\item{...}{Additional parameters passed to write.table().}
}
\description{
This doesn't record everything (e.g., bases), #' so using write.counts ->
read.counts is not guarenteed to be idempotent.  This just uses `countframe()`
to produce the data frame.
}
