% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandit.R
\name{AbstractBandit}
\alias{AbstractBandit}
\title{External AbstractBandit}
\description{
AbstractBandit intro
}
\section{Usage}{

\preformatted{b <- AbstractBandit$new()

b$reset()

print(b)
}
}

\section{Arguments}{

\describe{
\item{b}{A \code{AbstractBandit} object.}
}
}

\section{Details}{

\code{$new()} starts a new AbstractBandit, it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.
}

\examples{
\dontrun{}

}
\seealso{
Other contextual classes: \code{\link{Contextual}}
}
