% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{AbstractPolicy}
\alias{AbstractPolicy}
\title{Policy: Abstract Root Class}
\description{
The R6 class \code{Policy} is the root of all policies implemented by \code{\{contextual\}}.
That is, every \code{Policy} class in the \code{\{contextual\}} package has to inherit from, and implement the methods of,
the \code{Policy} superclass.
}
\section{Usage}{

\preformatted{
policy <- Policy()
}
}

\seealso{
Core contextual classes: \code{\link{Contextual}}, \code{\link{Simulator}},
\code{\link{Agent}}, \code{\link{History}}, \code{\link{Plot}}

Bandit classes: \code{\link{AbstractBandit}}, \code{\link{BasicBandit}},
\code{\link{LiLogBandit}}, \code{\link{SyntheticBandit}}

Other contextual policies: \code{\link{EpsilonFirstPolicy}},
  \code{\link{EpsilonGreedyPolicy}},
  \code{\link{Exp3Policy}}, \code{\link{LinUCBPolicy}},
  \code{\link{OraclePolicy}}, \code{\link{RandomPolicy}},
  \code{\link{ThompsonSamplingPolicy}}
}
