% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent_adversarial.R
\name{AdversarialAgent}
\alias{AdversarialAgent}
\title{Agent: Adversarial}
\description{
Agent intro
}
\section{Usage}{

\preformatted{b <- AdversarialAgent$new()

b$reset()

print(b)
}
}

\section{Arguments}{

\describe{
\item{b}{A \code{AdversarialAgent} object.}
}
}

\section{Details}{

\code{$new()} starts a new AdversarialAgent, it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.
}

\examples{
\dontrun{}

}
