% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agent.R
\name{Agent}
\alias{Agent}
\alias{do_step}
\alias{get_t}
\alias{set_t}
\title{Agent}
\description{
The R6 class \code{Agent} is responsible for the state, flow of information
between and the running of one \code{Bandit} and \code{Policy} pair.
As such, multiple \code{Agent}s can be run in parallel with each separate Agent keeping
track of \code{t} and the parameters in named list \code{theta} for its assigned \code{Policy}
and \code{Bandit} pair.
}
\section{Usage}{

\preformatted{
agent <- Agent$new(policy, bandit)
}
}

\section{Arguments}{


\describe{
\item{\code{policy}}{
A \code{Policy} object, expected to take into account the current \code{d} dimensional \code{context}
feature vector \code{X}, together with a limited set of parameters denoted \code{theta} (summarizing
all past actions), to choose one of the k arms of its corresponding bandit's arms at each time step \code{t}.
}
\item{\code{bandit}}{
A \code{Bandit} object, responsible for both the generation of \code{d} dimensional \code{context}
vectors \code{X} and the \code{k} I.I.D. distributions each generating a \code{reward} for each of
its \code{k} arms at each time step \code{t}.
}
}
}

\section{Methods}{


\describe{

\item{\code{set_t(t)}}{
Setter function, sets the state of the state variable holding the current time step \code{t}.
}

\item{\code{do_step()}}{
Completes one time step \code{t} by consecutively calling
\code{bandit$get_context()}, \code{policy$get_action()}, \code{bandit$get_reward()} and \code{policy$set_reward()}.
}

}
}

\seealso{
Core contextual classes: \code{\link{Bandit}}, \code{\link{Policy}}, \code{\link{Simulator}},
\code{\link{Agent}}, \code{\link{History}}, \code{\link{Plot}}

Bandit subclass examples: \code{\link{BasicBandit}}, \code{\link{ContextualBandit}},  \code{\link{LiSamplingOfflineBandit}}

Policy subclass examples: \code{\link{EpsilonGreedyPolicy}}, \code{\link{ContextualThompsonSamplingPolicy}}
}
