% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandit.R
\name{Bandit}
\alias{Bandit}
\alias{post_initialization}
\alias{get_context}
\alias{generate_bandit_data}
\title{Bandit}
\description{
R6 class \code{Bandit} is the parent of all \code{\{contextual\}} \code{Bandit} subclasses.
In \code{\{contextual\}}, \code{Bandits} are responsible for the generation of contexts and rewards.
}
\details{
On initialisation, any \code{Bandit} subclass has to define the number of arms \code{self$k}
and the number of contextual feature dimensions \code{self$d}.

\code{Bandit} subclasses are also responsible for returning \code{self$d x self$k} context matrices
through \code{get_context()}, and returning rewards through \code{get_reward()}.

\code{Bandit} subclasses have option to (pre-)generate these values in \code{generate_bandit_data()}.
}
\section{Usage}{

\preformatted{
  policy <- Bandit$new()
}
}

\section{Methods}{


\describe{

\item{\code{new()}}{ Generates and instantializes a new \code{Bandit} instance. }

\item{\code{get_context(t)}}{
argument:
\itemize{
\item \code{t}: integer, time step t.
}
returns a named list \code{list(k = n_arms, d = n_features, X = context)}
containing the current \code{d x k} dimensional \code{context} matrix \code{X},
the number of arms \code{k} and the number of features \code{d}.
}

\item{\code{get_reward(t, context, action)}}{
arguments:
\itemize{
\item \code{t}: integer, time step t.
\item \code{context}: list, containing \code{context$X} (d x k context matrix) \code{context$d}
(number of context feaures) and \code{context$k} (number of arms).
\item \code{action}:  list, containing \code{policy}'s \code{action$choice}.
}
returns:
\itemize{
\item \code{list} containing \code{reward} and, if available, the \code{optimal} (or "oracle")
reward for the current \code{t}
}
}

\item{\code{post_initialization()}}{
Called right after \code{Simulator} sets its seed,
but before it starts iterating over all time steps \code{t} in T.
If you need to set random values that remain available throughout the life of a \code{Bandit}
instance, this is the place to do so.
}

\item{\code{generate_bandit_data()}}{
Helper function called before \code{Simulator} starts iterating over all time steps \code{t} in T.
This function is only called when \code{bandit$precaching} is \code{TRUE} (default \code{FALSE}).
Pregenerate \code{contexts} and \code{rewards} here.
}
}
}

\seealso{
Core contextual classes: \code{\link{Bandit}}, \code{\link{Policy}}, \code{\link{Simulator}},
\code{\link{Agent}}, \code{\link{History}}, \code{\link{Plot}}

Bandit subclass examples: \code{\link{MabWeightBandit}}, \code{\link{ContextualBasicBandit}},  \code{\link{LiSamplingOfflineBandit}}

Policy subclass examples: \code{\link{EpsilonGreedyPolicy}}, \code{\link{ContextualThompsonSamplingPolicy}}
}
