% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandit.R
\name{Bandit}
\alias{Bandit}
\title{Bandit}
\description{
The R6 class \code{Bandit} is the parent class of all \code{Bandits} implemented in \code{\{contextual\}}.
Classes that extend the abstract superclass \code{Bandit} are responsible for both the generation
of \code{d} dimensional \code{context} vectors \code{X} and the \code{k} I.I.D. distributions
each generating a \code{reward} for each of its \code{k} arms at each time step \code{t}.
\code{Bandit} subclasses may (pre)generate these values synthetically, based on offline data, etc.
}
\section{Usage}{

\preformatted{
  policy <- Bandit$new()
}
}

\section{Methods}{


\describe{

\item{\code{new()}}{ Generates and initializes a new \code{Bandit} object. }

\item{\code{pre_calculate()}}{
Called right after \code{Simulator} sets its seed,
but before it starts iterating over all time steps \code{t} in T. If you need to initialize random values in a \code{Policy},
this is the place to do so.
}

\item{\code{get_context(t)}}{
Returns a named list \code{list(k = n_arms, d = n_features, X = context)}
with the current \code{d} dimensional \code{context} feature vector \code{X} together with the number of arms \code{k}.
}

\item{\code{get_reward(t, context, action)}}{
Returns the named list \code{list(reward = reward_for_choice_made,
     optimal = optimal_reward_value)} containing the \code{reward}
for the \code{action} previously returned by \code{policy} and,
optionally, the \code{optimal} reward
at the current time \code{t}.
}

\item{\code{generate_bandit_data()}}{
A helper function that is called before \code{Simulator} starts iterating over all time steps \code{t} in T.
This function is called when \code{bandit$precaching} has been set to \code{TRUE}.
Pregenerate \code{contexts} and \code{rewards} here.
}
}
}

\seealso{
Core contextual classes: \code{\link{Simulator}},
\code{\link{Agent}}, \code{\link{History}}, \code{\link{Plot}}

Bandit classes: \code{\link{Bandit}}, \code{\link{BasicBandit}},
\code{\link{LiSamplingOfflineBandit}}, \code{\link{SyntheticBandit}}

Other contextual bandits: \code{\link{DoublyRobustOfflineBandit}},
  \code{\link{LiSamplingOfflineBandit}}
}
