% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandit.R
\name{Bandit}
\alias{Bandit}
\title{External Bandit}
\description{
Bandit intro
}
\section{Usage}{

\preformatted{b <- Bandit$new()

b$reset()

print(b)
}
}

\section{Arguments}{

\describe{
\item{b}{A \code{Bandit} object.}
}
}

\section{Details}{

\code{$new()} starts a new Bandit, it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.
}

\examples{
\dontrun{}

}
