% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_mab_btslogit.R
\name{BTSLogitPolicy}
\alias{BTSLogitPolicy}
\title{Policy: BTSLogitPolicy}
\description{
Each time step t, \code{BTS} runs a linear regression that produces coefficients for each context feature \code{d}.
It then observes the new context, and generates a predicted payoff or reward together with a confidence interval for each available arm.
It then proceeds to choose the arm with the highest upper confidence bound.
}
\section{Usage}{

\preformatted{
policy <- BTSLogitPolicy(alpha = 1.0, name = "LinUCB")
}
}

\seealso{
Core contextual classes: \code{\link{Bandit}}, \code{\link{Policy}}, \code{\link{Simulator}},
\code{\link{Agent}}, \code{\link{History}}, \code{\link{Plot}}

Other contextual subclasses: \code{\link{BasicBandit}},
  \code{\link{ContextualBandit}},
  \code{\link{ContextualThompsonSamplingPolicy}},
  \code{\link{ContinuumBandit}},
  \code{\link{DoublyRobustOfflineBandit}},
  \code{\link{EpsilonFirstPolicy}},
  \code{\link{EpsilonGreedyPolicy}},
  \code{\link{Exp3Policy}},
  \code{\link{LiSamplingOfflineBandit}},
  \code{\link{LifPolicy}},
  \code{\link{LinUCBDisjointPolicy}},
  \code{\link{LinUCBHybridPolicy}},
  \code{\link{OraclePolicy}}, \code{\link{RandomPolicy}},
  \code{\link{SoftmaxPolicy}},
  \code{\link{SyntheticBandit}},
  \code{\link{ThompsonSamplingPolicy}},
  \code{\link{UCB1Policy}}
}
