% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-output.R
\name{check.config}
\alias{check.config}
\title{Check for obvious errors in a configuration.}
\usage{
check.config(config)
}
\arguments{
\item{config}{A configuration.}
}
\value{
TRUE or FALSE
}
\description{
Check for obvious errors in a configuration.
}
