% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residuals.R
\name{computeresids}
\alias{computeresids}
\title{Compute Residuals}
\usage{
computeresids(model, pretty = TRUE, in_longwin = longwin(model),
  in_shortwin = shortwin(model), in_leftwin = leftwin(model),
  counts = NULL, genmatrixfile = NULL, overlapping = FALSE)
}
\description{
Compute Residuals
}
