% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_matrix.R
\name{computetransmatrix}
\alias{computetransmatrix}
\title{Find the Transition Matrix}
\usage{
computetransmatrix(genmatrix, projmatrix, tlen = 1, shape = 1,
  names = FALSE, transpose = FALSE, time = "fixed", ...)
}
\description{
Compute the product of exp(tlen*genmatrix) and projmatrix, either on the left or the right (as transpose is true or false)
  either after a fixed time: exp(tlen*genmatrix)
  or after a gamma-distributed time
}
