% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandit_contextual.R
\name{ContextualBandit}
\alias{ContextualBandit}
\title{Bandit: ContextualBandit}
\description{
ContextualBandit intro
}
\section{Usage}{

\preformatted{b <- ContextualBandit$new()

b$reset()

print(b)
}
}

\section{Arguments}{

\describe{
\item{b}{A \code{ContextualBandit} object.}
}
}

\section{Details}{

\code{$new()} starts a new ContextualBandit, it uses \code{\link[base]{pipe}}.
R does \emph{not} wait for the process to finish, but returns
immediately.
}

\examples{
\dontrun{}

}
