% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_cmab_epoch_greedy.R
\name{ContextualEpochGreedyDisjointPolicy}
\alias{ContextualEpochGreedyDisjointPolicy}
\title{Policy: A Time and Space Efficient Algorithm for Contextual Linear Bandits}
\description{
Policy: A Time and Space Efficient Algorithm for Contextual Linear Bandits
}
\section{Usage}{

\preformatted{
policy <- EpsilonGreedyPolicy(epsilon = 0.1)
}
}

\seealso{
Other contextual subclasses: \code{\link{BasicBandit}},
  \code{\link{ContextualThompsonSamplingPolicy}},
  \code{\link{ContinuumBandit}},
  \code{\link{EpsilonFirstPolicy}},
  \code{\link{EpsilonGreedyPolicy}},
  \code{\link{Exp3Policy}},
  \code{\link{GittinsBrezziLaiPolicy}},
  \code{\link{LiSamplingOfflineBandit}},
  \code{\link{LifPolicy}},
  \code{\link{LinUCBDisjointOptimizedPolicy}},
  \code{\link{LinUCBDisjointPolicy}},
  \code{\link{LinUCBGeneralPolicy}},
  \code{\link{LinUCBHybridOptimizedPolicy}},
  \code{\link{LinUCBHybridPolicy}},
  \code{\link{LinUCBOFULPolicy}},
  \code{\link{OraclePolicy}}, \code{\link{RandomPolicy}},
  \code{\link{SoftmaxPolicy}},
  \code{\link{SyntheticBandit}},
  \code{\link{ThompsonSamplingPolicy}},
  \code{\link{UCB1Policy}}
}
