% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_cmab_bts_logit.R
\name{ContextualLogitBTSPolicy}
\alias{ContextualLogitBTSPolicy}
\title{Policy: ContextualLogitBTSPolicy}
\description{
Each time step t, \code{ContextualLogitBTSPolicy} runs ...
}
\section{Usage}{

\preformatted{
policy <- ContextualLogitBTSPolicy()
}
}

\seealso{
Core contextual classes: \code{\link{Bandit}}, \code{\link{Policy}}, \code{\link{Simulator}},
\code{\link{Agent}}, \code{\link{History}}, \code{\link{Plot}}

Bandit subclass examples: \code{\link{BasicBernoulliBandit}}, \code{\link{ContextualLogitBandit}},  \code{\link{OfflinePolicyEvaluatorBandit}}

Policy subclass examples: \code{\link{EpsilonGreedyPolicy}}, \code{\link{ContextualThompsonSamplingPolicy}}
}
