% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cx_stop.r
\name{cx_stop}
\alias{cx_stop}
\title{Stop with formatted error message}
\usage{
cx_stop(message, ..., .envir = parent.frame(), .call_level = -1)
}
\arguments{
\item{message}{Main error message}

\item{...}{Additional message components using cli syntax}

\item{.envir}{Environment for string interpolation}

\item{.call_level}{How far back in the call stack to look for the function name (default -1)}
}
\description{
Stop with formatted error message
}
