% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{get_arm_context}
\alias{get_arm_context}
\title{Returns context of arm}
\usage{
get_arm_context(X, arm, select_features = NULL)
}
\arguments{
\item{X}{d x k Matrix}

\item{arm}{index of arm}
}
\value{
Vector that represents context related to an arm
}
\description{
Given d x k matrix or d dimensional vector X,
returns a vector with arm's context.
}
