% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{get_full_context}
\alias{get_full_context}
\title{Get full context matrix over all arms}
\usage{
get_full_context(context, select_features = NULL, prepend_arm_matrix = FALSE)
}
\arguments{
\item{context}{a context list containing a d x k Matrix or
d dimensional context vector X, the number of features d and
number of arms k.}

\item{select_features}{indices of to be returned feature rows.b}

\item{prepend_arm_matrix}{prepend a diagonal arm matrix to the returned context vector. That is,
when k = 5 arms, prepend diag(5) to the top of the matrix.}
}
\value{
A d x k context Matrix
}
\description{
Given matrix or d dimensional vector X,
number of arms k and number of features d
returns a matrix with d x k context matrix
}
