% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_oracle.R
\name{OraclePolicy}
\alias{OraclePolicy}
\title{Policy: Oracle}
\description{
\code{OraclePolicy} is a also known as a "cheating" or "godlike"
policy, as it knows the reward probabilities at all times,
and will always play the optimal arm. It is often used as
a baseline to compare other policies to.
}
\section{Usage}{

\preformatted{
policy <- OraclePolicy(name = "OraclePolicy")
}
}

\section{Arguments}{


\describe{
\item{\code{name}}{
character string specifying this policy. \code{name}
is, amongst others, saved to the History log and displayed in summaries and plots.
}
}
}

\section{Methods}{


\describe{
\item{\code{new(name = "Oracle")}}{ Generates a new \code{OraclePolicy} object. Arguments are defined in the Argument section above.}
}

\describe{
\item{\code{set_parameters()}}{each policy needs to assign the parameters it wants to keep track of
to list \code{self$parameters} that has to be defined in \code{set_parameters()}'s body.
The parameters defined here can later be accessed by arm index in the following way:
\code{theta[[index_of_arm]]$parameter_name}
}
}

\describe{
\item{\code{get_action(context)}}{
here, a policy decides which arm to choose, based on the current values
of its parameters and, potentially, the current context.
}
}

\describe{
\item{\code{set_reward(reward, context)}}{
in \code{set_reward(reward, context)}, a policy updates its parameter values
based on the reward received, and, potentially, the current context.
}
}
}

\examples{

horizon            <- 100L
simulations        <- 100L
weight_per_arm     <- c(0.9, 0.1, 0.1)

policy             <- OraclePolicy$new(name = "Oracle")
bandit             <- SyntheticBandit$new(weights = weight_per_arm, precache = FALSE)
agent              <- Agent$new(policy, bandit)

history            <- Simulator$new(agent, horizon, simulations, do_parallel = FALSE)$run()

plot(history, type = "grid")


}
\references{
Gittins, J., Glazebrook, K., & Weber, R. (2011). Multi-armed bandit allocation indices. John Wiley & Sons. (Original work published 1989)
}
\seealso{
Core contextual classes: \code{\link{Contextual}}, \code{\link{Simulator}},
\code{\link{Agent}}, \code{\link{History}}, \code{\link{Plot}}

Bandit classes: \code{\link{AbstractBandit}}, \code{\link{BasicBandit}},
\code{\link{OfflineLiBandit}}, \code{\link{SyntheticBandit}}

Other contextual policies: \code{\link{AbstractPolicy}},
  \code{\link{EpsilonFirstPolicy}},
  \code{\link{EpsilonGreedyPolicy}},
  \code{\link{Exp3Policy}}, \code{\link{LinUCBPolicy}},
  \code{\link{RandomPolicy}},
  \code{\link{ThompsonSamplingPolicy}}
}
