% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-output.R
\name{parse.models}
\alias{parse.models}
\title{Parse models.}
\usage{
parse.models(config, do.fixfns = TRUE)
}
\value{
A model configuration.
}
\description{
Check that all models are specified,
 fill in defaults,
 turn fixfn into functions, etc.
}
