% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_matrix.R
\name{peel.transmat}
\alias{peel.transmat}
\title{Compute probabilities of all combinations, on a tree}
\usage{
peel.transmat(tree, rowtaxon, coltaxa, models, genmatrices, projmatrix,
  root.distrn, tlens = tree$edge.length, return.list = TRUE,
  debug = FALSE)
}
\arguments{
\item{rowtaxon}{The name of a single tip, or the root}

\item{coltaxa}{A vector of names of tips}

\item{models}{A list of model names, in (tip,node) order}

\item{genmatrices}{A list of genmatrix'es whose names are model names}
}
\description{
Compute probabilities of all combinations, on a tree
}
