% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Plot}
\alias{Plot}
\alias{average}
\alias{optimal}
\alias{arms}
\alias{do_plot}
\alias{gg_color_hue}
\alias{check_history_data}
\title{Plot}
\description{
The R6 class \code{Plot} generates plots from on \code{History} data.
It is usually actually invoked by calling the generic \code{plot(h)}, where \code{h}
is an \code{History} class instance.
}
\section{Usage}{

\preformatted{
Plot <- Plot$new()
}
}

\section{Methods}{


\describe{

\item{\code{cumulative(history,...)}}{
Writes a plot of cumulative regret or reward (depending on parameter regret=TRUE/FALSE)
over time.
}
\item{\code{average(history,...)}}{
Writes a plot of average regret or reward (depending on parameter regret=TRUE/FALSE)
over time.
}
\item{\code{arms(history),...}}{
Writes a plot of the ratio with which each arm of a simulation was chosen over time.
If multiple agents are run, only plots the arm plot of the first agent.
}
}
}

\section{Plot method arguments}{


\describe{
\item{\code{regret}}{
\code{(logical, TRUE)} If regret is TRUE, regret will be plotted on the y-axis.
When regret is set to FALSE, reward will be plotted on the y-axis.
}
\item{\code{rate}}{
(\code{logical, TRUE)} If rate is TRUE, the rate of the regret or reward is plotted.
}
\item{\code{limit_agents}}{
\code{(list , NULL)} Limit the plotted agents to the agents in list.
}
\item{\code{no_par}}{
\code{(logical, FALSE)} If no_par is TRUE, Plot() does not set or adjust plotting parameters.
This makes it possible to set your own plotting parameters, for instance, when position multiple
Plots in a no_par.
}
\item{\code{legend}}{
\code{(logical, TRUE)} Print the legend.
}
\item{\code{legend_title}}{
\code{(character , NULL)} Set a legend title.
}
\item{\code{legend_labels}}{
\code{(list , NULL)} Set legend labels to custom values as specificed in this list.
}
\item{\code{legend_border}}{
\code{(logical , NULL)} Set a legend title.
}
\item{\code{xlim}}{
\code{(c(integer,integer), NULL)} Set x-axis limits.
}
\item{\code{ylim}}{
\code{(c(integer,integer), NULL)} Set y-axis limits.
}
\item{\code{use_colors}}{
\code{(logical, TRUE)} If use_colors is set to FALSE, plots will be in grayscale.
Otherwise, plots will make use of a color palette.
}
\item{\code{ci}}{
\code{(logical, FALSE)} When ci is TRUE, Plot() will display 95% confidence intervals.
}
\item{\code{traces}}{
\code{(logical , FALSE)} When traces is TRUE, Plot() will plot the traces of each independent simulation.
}
\item{\code{traces_max}}{
\code{(integer , 100)} The maximum number of trace lines.
}
\item{\code{traces_alpha}}{
\code{(numeric , 0.3)} Sets the opacity of all trace lines. By default traces_alpha is set to 0.3,
that is, at an opacity of 30%.
}
\item{\code{smooth}}{
\code{(logical , FALSE)} When smooth is TRUE, Plot() will smooth all plots.
}
\item{\code{interval}}{
\code{(integer, NULL)} Plot only for every every t%%interval==0
}
\item{\code{color_step}}{
\code{(integer, 1)} Linecharts will cycle through agents/color_step colors.
}
\item{\code{lty_step}}{
\code{(integer, 1)} Linecharts will cycle through agents/lty_step line types
}
\item{\code{lwd}}{
\code{(integer, 1)} Linecharts will be of lwd width.
}
}
}

\seealso{
Core contextual classes: \code{\link{Bandit}}, \code{\link{Policy}}, \code{\link{Simulator}},
\code{\link{Agent}}, \code{\link{History}}, \code{\link{Plot}}

Bandit subclass examples: \code{\link{BasicBernoulliBandit}}, \code{\link{ContextualLogitBandit}},  \code{\link{OfflinePolicyEvaluatorBandit}}

Policy subclass examples: \code{\link{EpsilonGreedyPolicy}}, \code{\link{ContextualThompsonSamplingPolicy}}
}
