% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{Plot}
\alias{Plot}
\title{Plot}
\description{
The R6 class \code{Plot} generates plots based on \code{History} data.
It is usually actually invoked by calling the generic \code{plot(h)}, where \code{h}
is an \code{History} class instance.
}
\section{Usage}{

\preformatted{
Plot <- Plot$new()
}
}

\section{Methods}{


\describe{

\item{\code{cumulative(history)}}{
Returns a plot of cumulative regret or reward over time.
}
\item{\code{average(history)}}{
Returns a plot of average regret or reward over time.
}
\item{\code{optimal(history)}}{
Returns a plot of the percentage the optimal arm was chosen over time.
}
\item{\code{arms(history)}}{
Returns a plot of the ratio with which each arm was chosen over time.
}

}
}

\seealso{
Core contextual classes: \code{\link{Simulator}},
\code{\link{Agent}}, \code{\link{History}}, \code{\link{Plot}}

Bandit classes: \code{\link{Bandit}}, \code{\link{BasicBandit}},
\code{\link{LiSamplingOfflineBandit}}, \code{\link{SyntheticBandit}}

Other contextual: \code{\link{Agent}},
  \code{\link{History}}, \code{\link{Simulator}},
  \code{\link{plot.history}}, \code{\link{print.history}},
  \code{\link{summary.history}}
}
