% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy.R
\name{Policy}
\alias{Policy}
\alias{get_action}
\alias{set_reward}
\alias{set_parameters}
\alias{initialize_theta}
\title{Policy}
\description{
R6 class \code{Policy} is the parent class of all \code{Policy} implementations in
\code{\{contextual\}}. Classes that extend this abstract superclass are expected to take into
account the current \code{d} dimensional \code{context},
together with a limited set of parameters denoted \code{theta} (summarizing all past actions),
to choose one of \code{k} \code{Bandit} arms at each time step \code{t}.
On choosing one of the \code{k} arms and receiving its corresponding \code{reward},
the \code{Policy} then utilizes the current \code{context}, \code{action} and \code{reward} to
update its set of parameters \code{theta}.
}
\section{Usage}{

\preformatted{
  policy <- Policy$new()
}
}

\section{Methods}{


\describe{
\item{\code{new()}}{
Generates and initializes a new \code{Policy} object.
}

\item{\code{get_action(t, context)}}{
arguments:
\itemize{
\item \code{t}: integer, time step \code{t}.
\item \code{context}: list, containing the current \code{context$X} (d x k context matrix),
\code{context$k} (number of arms) and \code{context$d} (number of context feaures)
}
computes which arm to play based on the current values in named list \code{theta}
and the current \code{context}. Returns a named list containing
\code{action$choice}, which holds the index of the arm to play.
}

\item{\code{set_reward(t, context, action, reward)}}{
arguments:
\itemize{
\item \code{t}: integer, time step \code{t}.
\item \code{context}: list, containing the current \code{context$X} (d x k context matrix),
\code{context$k} (number of arms) and \code{context$d} (number of context feaures)
(as set by \code{bandit}).
\item \code{action}:  list, containing \code{action$choice} (as set by \code{policy}).
\item \code{reward}:  list, containing \code{reward$reward} and, if available,
\code{reward$optimal} (as set by \code{bandit}).
}
utilizes the above arguments to update and return the set of parameters in list \code{theta}.
}

\item{\code{set_parameters()}}{
Helper function, called during a Policy's initialisation, assigns the values
it finds in list \code{self$theta_to_arms} to each of the Policy's k arms.
The parameters defined here can then be accessed by arm index in the following way:
\code{theta[[index_of_arm]]$parameter_name}.
}

}
}

\seealso{
Core contextual classes: \code{\link{Bandit}}, \code{\link{Policy}}, \code{\link{Simulator}},
\code{\link{Agent}}, \code{\link{History}}, \code{\link{Plot}}

Bandit subclass examples: \code{\link{MabWeightBandit}}, \code{\link{ContextualLogitBandit}},  \code{\link{LiSamplingOfflineBandit}}

Policy subclass examples: \code{\link{EpsilonGreedyPolicy}}, \code{\link{ContextualThompsonSamplingPolicy}}
}
