% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_counts.R
\name{predictcounts}
\alias{predictcounts}
\title{Predict Counts from a Model}
\usage{
predictcounts(longwin, shortwin, leftwin, initcounts, mutrates, selcoef,
  genmatrix, projmatrix, params = NULL, tlen = 1)
}
\description{
Compute expected counts of paired patterns.
}
