% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_generic.R
\name{print.history}
\alias{print.history}
\alias{print.History}
\title{Print Method for Contextual History}
\usage{
\method{print}{History}(x, ...)
}
\arguments{
\item{x}{A \code{History} object.}

\item{...}{Further plotting parameters.}
}
\description{
print.history, a method for the print generic. It is designed for a quick look at History data.
}
\seealso{
Core contextual classes: \code{\link{Simulator}},
\code{\link{Agent}}, \code{\link{History}}, \code{\link{Plot}}

Bandit classes: \code{\link{Bandit}}, \code{\link{BasicBernoulliBandit}},
\code{\link{OfflinePolicyEvaluatorBandit}}, \code{\link{ContextualLogitBandit}}
}
