% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_counts.R
\name{projectcounts}
\alias{projectcounts}
\title{Compute Counts for Shorter Tmers.}
\usage{
projectcounts(counts, new.leftwin = leftwin(counts),
  new.shortwin = shortwin(counts), new.longwin = longwin(counts),
  overlapping = FALSE)
}
\description{
Valid ranges for parameters are
   (l-lc)^+ <= k < (l+w)-(lc+wc)+(r-rc)^-
  where 
      l = leftwin, lc = new.leftwin
      w = shortwin, wc = new.shortwin
      r = rightwin, rc = new.rightwin
If the original counts were from overlapping windows, then this will overcount the resulting patterns:
   if you slide a window of length L in steps of size 1 then a subwindow of size W
     will be seen in ( (L-W) ) big windows;
   so we need to divide the counts by the factor 'overcount' below
     ... but patterns at the boundary of the sequence will not be overcounted. 
    Take the ceiling of the resulting counts to fix these.
}
\examples{
counts <- counttrans(ipatterns=getpatterns(3,bases=c("X","O")), fpatterns=getpatterns(2,bases=c("X","O")), 
                        initseq="XOOXXXOXOXOXOOOXXO", finalseq="XOXOXOXOXXOXOOOXXO", leftwin=1)
projectcounts(counts, new.longwin=2, new.leftwin=0) 
projectcounts(counts, new.shortwin=1)

}
