% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-output.R
\name{read.config}
\alias{read.config}
\title{Read in JSON Config File}
\usage{
read.config(configfile, quiet = FALSE, json)
}
\value{
A model configuration (list).
}
\description{
... or directly from a text string. Automatically fills in zero-length
parameters, and fills in selfactors if it isn't there, or moves values from
selpats if it is.
}
