% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input-output.R
\name{read.counts}
\alias{read.counts}
\title{Read in a File of Counts}
\usage{
read.counts(infile, leftwin, bases, longpats, shortpats, skip = 0)
}
\arguments{
\item{infile}{File name.}

\item{leftwin}{The T-mer's left overhang (default: read from leading configuration line).}

\item{bases}{The alphabet (default: unique characters seen in first column).}

\item{longpats}{Patterns indexing rows in the result (default: all n-mers).}

\item{shortpats}{Patterns indexing columns in the result (default: all k-mers).}

\item{skip}{Number of lines at the top to skip (default: the header).}
}
\value{
A tuplecounts object.
}
\description{
The file of counts should be of the following form (including the leading commented line):
}
\details{
# { "leftwin" : 3 }
       taxon1  taxon2 ... count
        AAAA      AA ...    19
        CAAA      AA ...     2
        GAAA      AA ...     6
        TAAA      AA ...     3

... and convert it to a 'tuplecounts' object
optionally passing in the orderings of the rows and columns

If leftwin is not specified it will be looked for *in json format* on the
first line of the file after '#' ... mind the braces and quotes!
}
