% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees.R
\name{reorder.counts}
\alias{reorder.counts}
\title{Reorder Counts by Taxon Reordering}
\usage{
\method{reorder}{counts}(counts, new.ord)
}
\arguments{
\item{counts}{A tuplecounts object.}

\item{new.ord}{A vector of taxon names, a reordering of coltaxa(counts)}
}
\value{
A new tuplecounts object.
}
\description{
Rearrange the columns of counts to a new ordering of taxa
}
