% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{sim_post}
\alias{sim_post}
\title{Binomial Posterior Simulator}
\usage{
sim_post(x, n, alpha = 1, beta = 1, ndraws = 5000)
}
\arguments{
\item{x}{Vector of the number of successes per arm.}

\item{n}{Vector of the number of trials per arm.}

\item{alpha}{a positive-definite symmetric matrix specifying the covariance matrix of the variables.}

\item{beta}{Shape parameter beta for the prior beta distribution.}

\item{ndraws}{Number of random draws from the posterior.}

\item{alpha}{Shape parameter alpha for the prior beta distribution.}
}
\value{
Matrix of bayesian probabilities for each arm being the best binomial bandit
}
\description{
Simulates the posterior distribution of
the Bayesian probabilities for each arm being the
best binomial bandit.
}
\examples{

n=c(100,102,120,130)
sim_post(x,n)

}
\author{
Thomas Lotze and Markus Loecher
}
