% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_mab_simplebts.R
\name{SimpleBTSPolicy}
\alias{SimpleBTSPolicy}
\title{Policy: SimpleBTSPolicy}
\description{
Each time step t, \code{SimpleBTSPolicy} runs ...
}
\section{Usage}{

\preformatted{
policy <- SimpleBTSPolicy(alpha = 1.0)
}
}

\seealso{
Core contextual classes: \code{\link{Bandit}}, \code{\link{Policy}}, \code{\link{Simulator}},
\code{\link{Agent}}, \code{\link{History}}, \code{\link{Plot}}

Bandit subclass examples: \code{\link{BasicBernoulliBandit}}, \code{\link{ContextualLogitBandit}},  \code{\link{OfflinePolicyEvaluatorBandit}}

Policy subclass examples: \code{\link{EpsilonGreedyPolicy}}, \code{\link{ContextualThompsonSamplingPolicy}}
}
