% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_mab_simplebts.R
\name{SimpleBTSPolicy}
\alias{SimpleBTSPolicy}
\title{Policy: SimpleBTSPolicy}
\description{
Each time step t, \code{SimpleBTSPolicy} runs ...
}
\section{Usage}{

\preformatted{
policy <- SimpleBTSPolicy(alpha = 1.0, name = "LinUCB")
}
}

\seealso{
Core contextual classes: \code{\link{Bandit}}, \code{\link{Policy}}, \code{\link{Simulator}},
\code{\link{Agent}}, \code{\link{History}}, \code{\link{Plot}}
}
