% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/policy_mab_simplebts.R
\name{SimpleBTSPolicy}
\alias{SimpleBTSPolicy}
\title{Policy: SimpleBTSPolicy}
\description{
Each time step t, \code{SimpleBTSPolicy} runs ...
}
\section{Usage}{

\preformatted{
policy <- SimpleBTSPolicy(alpha = 1.0, name = "LinUCB")
}
}

\seealso{
Other contextual policies: \code{\link{AbstractPolicy}},
  \code{\link{BTSLogitPolicy}},
  \code{\link{ContextualEpsilonGreedyPolicy}},
  \code{\link{ContextualThompsonSamplingPolicy}},
  \code{\link{EpsilonFirstPolicy}},
  \code{\link{EpsilonGreedyPolicy}},
  \code{\link{Exp3Policy}}, \code{\link{LinUCBPolicy}},
  \code{\link{OraclePolicy}}, \code{\link{RandomPolicy}},
  \code{\link{SoftmaxPolicy}},
  \code{\link{ThompsonSamplingPolicy}},
  \code{\link{UCB1Policy}}
}
