% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_generic.R
\name{summary.history}
\alias{summary.history}
\alias{summary.History}
\title{Summary Method for Contextual History}
\usage{
\method{summary}{History}(object, ...)
}
\arguments{
\item{object}{A \code{History} object.}

\item{...}{Further summary parameters.}
}
\description{
summary.history, a method for the summary generic. It is designed for a quick summary of History data.
}
\seealso{
Core contextual classes: \code{\link{Simulator}},
\code{\link{Agent}}, \code{\link{History}}, \code{\link{Plot}}

Bandit classes: \code{\link{Bandit}}, \code{\link{BasicBandit}},
\code{\link{LiSamplingOfflineBandit}}, \code{\link{SyntheticBandit}}

Other contextual: \code{\link{plot.history}},
  \code{\link{print.history}}
}
