% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions_utility.R
\name{which_max_tied}
\alias{which_max_tied}
\title{Get maximum value}
\usage{
which_max_tied(vec)
}
\arguments{
\item{vec}{vector of values}
}
\description{
Returns the index of the maximum value in vector \code{vec}.
}
\details{
If there is a tie,
the index of one of the tied maxima is returned at random.
}
